hopr_chain_types

Module chain_events

Types that are related to events that are raised on-chain and extracted from chain logs.

These events happen in response to actions (transactions, smart contract calls) done by a HOPR node on chain.

See chain-actions and chain-indexer crates for details.

Structs

SignificantChainEvent

Contains TX hash along with the Chain Event data. This could be used to pair up some events with Action. Each Action is typically concluded by a significant chain event.

Enums

ChainEventType

Enumeration of HOPR chain events.

NetworkRegistryStatus

Status of a node in network registry.