hopr_chain_types::chain_events

Enum ChainEvent

pub enum ChainEvent {
    Announcement(AccountEntry),
    ChannelOpened(ChannelEntry),
    ChannelClosureInitiated(ChannelEntry),
    ChannelClosed(ChannelEntry),
    ChannelBalanceIncreased(ChannelEntry, HoprBalance),
    ChannelBalanceDecreased(ChannelEntry, HoprBalance),
    TicketRedeemed(ChannelEntry, Option<Box<VerifiedTicket>>),
    WinningProbabilityIncreased(WinningProbability),
    WinningProbabilityDecreased(WinningProbability),
    TicketPriceChanged(HoprBalance),
}

Enumeration of HOPR chain events.

Variants

Announcement(AccountEntry)

Peer on-chain announcement event.

The AccountEntry is guaranteed to be announced.

ChannelOpened(ChannelEntry)

A new channel has been opened

The ChannelEntry is guaranteed to be opened.

ChannelClosureInitiated(ChannelEntry)

Channel closure has been initiated.

The ChannelEntry is guaranteed to be pending to close.

ChannelClosed(ChannelEntry)

Channel closure has been finalized.

The ChannelEntry is guaranteed to be closed.

ChannelBalanceIncreased(ChannelEntry, HoprBalance)

Channel balance has increased by an amount.

The [HoprBalance] is never 0 and represents the difference from the current new balance on the ChannelEntry.

ChannelBalanceDecreased(ChannelEntry, HoprBalance)

Channel balance has decreased by an amount.

The [HoprBalance] is never 0 and represents the difference from the current new balance on the ChannelEntry.

TicketRedeemed(ChannelEntry, Option<Box<VerifiedTicket>>)

Ticket has been redeemed on a channel.

If the channel is a node’s own, it also contains the ticket that has been redeemed.

WinningProbabilityIncreased(WinningProbability)

The minimum winning probability has been increased.

WinningProbabilityDecreased(WinningProbability)

The minimum winning probability has been decreased.

TicketPriceChanged(HoprBalance)

A new ticket price has been set.

Implementations

impl ChainEvent

pub fn try_as_announcement(self) -> Option<AccountEntry>

pub const fn try_as_announcement_ref(&self) -> Option<&AccountEntry>

pub fn try_as_announcement_mut(&mut self) -> Option<&mut AccountEntry>

pub fn try_as_channel_opened(self) -> Option<ChannelEntry>

pub const fn try_as_channel_opened_ref(&self) -> Option<&ChannelEntry>

pub fn try_as_channel_opened_mut(&mut self) -> Option<&mut ChannelEntry>

pub fn try_as_channel_closure_initiated(self) -> Option<ChannelEntry>

pub const fn try_as_channel_closure_initiated_ref( &self, ) -> Option<&ChannelEntry>

pub fn try_as_channel_closure_initiated_mut( &mut self, ) -> Option<&mut ChannelEntry>

pub fn try_as_channel_closed(self) -> Option<ChannelEntry>

pub const fn try_as_channel_closed_ref(&self) -> Option<&ChannelEntry>

pub fn try_as_channel_closed_mut(&mut self) -> Option<&mut ChannelEntry>

pub fn try_as_channel_balance_increased( self, ) -> Option<(ChannelEntry, HoprBalance)>

pub const fn try_as_channel_balance_increased_ref( &self, ) -> Option<(&ChannelEntry, &HoprBalance)>

pub fn try_as_channel_balance_increased_mut( &mut self, ) -> Option<(&mut ChannelEntry, &mut HoprBalance)>

pub fn try_as_channel_balance_decreased( self, ) -> Option<(ChannelEntry, HoprBalance)>

pub const fn try_as_channel_balance_decreased_ref( &self, ) -> Option<(&ChannelEntry, &HoprBalance)>

pub fn try_as_channel_balance_decreased_mut( &mut self, ) -> Option<(&mut ChannelEntry, &mut HoprBalance)>

pub fn try_as_ticket_redeemed( self, ) -> Option<(ChannelEntry, Option<Box<VerifiedTicket>>)>

pub const fn try_as_ticket_redeemed_ref( &self, ) -> Option<(&ChannelEntry, &Option<Box<VerifiedTicket>>)>

pub fn try_as_ticket_redeemed_mut( &mut self, ) -> Option<(&mut ChannelEntry, &mut Option<Box<VerifiedTicket>>)>

pub fn try_as_winning_probability_increased(self) -> Option<WinningProbability>

pub const fn try_as_winning_probability_increased_ref( &self, ) -> Option<&WinningProbability>

pub fn try_as_winning_probability_increased_mut( &mut self, ) -> Option<&mut WinningProbability>

pub fn try_as_winning_probability_decreased(self) -> Option<WinningProbability>

pub const fn try_as_winning_probability_decreased_ref( &self, ) -> Option<&WinningProbability>

pub fn try_as_winning_probability_decreased_mut( &mut self, ) -> Option<&mut WinningProbability>

pub fn try_as_ticket_price_changed(self) -> Option<HoprBalance>

pub const fn try_as_ticket_price_changed_ref(&self) -> Option<&HoprBalance>

pub fn try_as_ticket_price_changed_mut(&mut self) -> Option<&mut HoprBalance>

impl ChainEvent

pub const fn is_announcement(&self) -> bool

Returns true if the enum is ChainEvent::Announcement otherwise false

pub const fn is_channel_opened(&self) -> bool

Returns true if the enum is ChainEvent::ChannelOpened otherwise false

pub const fn is_channel_closure_initiated(&self) -> bool

Returns true if the enum is ChainEvent::ChannelClosureInitiated otherwise false

pub const fn is_channel_closed(&self) -> bool

Returns true if the enum is ChainEvent::ChannelClosed otherwise false

pub const fn is_channel_balance_increased(&self) -> bool

Returns true if the enum is ChainEvent::ChannelBalanceIncreased otherwise false

pub const fn is_channel_balance_decreased(&self) -> bool

Returns true if the enum is ChainEvent::ChannelBalanceDecreased otherwise false

pub const fn is_ticket_redeemed(&self) -> bool

Returns true if the enum is ChainEvent::TicketRedeemed otherwise false

pub const fn is_winning_probability_increased(&self) -> bool

Returns true if the enum is ChainEvent::WinningProbabilityIncreased otherwise false

pub const fn is_winning_probability_decreased(&self) -> bool

Returns true if the enum is ChainEvent::WinningProbabilityDecreased otherwise false

pub const fn is_ticket_price_changed(&self) -> bool

Returns true if the enum is ChainEvent::TicketPriceChanged otherwise false