hopr_chain_types::exports::alloy::consensus::proofs

Function state_root_unhashed

pub fn state_root_unhashed<A>(
    state: impl IntoIterator<Item = (Address, A)>,
) -> FixedBytes<32>
where
    A: Into<TrieAccount>,

Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.