hopr_chain_types::exports::alloy::consensus

Struct Block

pub struct Block<T, H = Header> {
    pub header: H,
    pub body: BlockBody<T, H>,
}

Ethereum full block.

Withdrawals can be optionally included at the end of the RLP encoded message.

Taken from reth-primitives

See p2p block encoding reference: https://github.com/ethereum/devp2p/blob/master/caps/eth.md#block-encoding-and-validity

Fields
header: H

Block header.

body: BlockBody<T, H>

Block body.

Implementations

impl<T, H> Block<T, H>

pub const fn new(header: H, body: BlockBody<T, H>) -> Block<T, H>

Creates a new block with the given header and body.

pub fn uncle(header: H) -> Block<T, H>

Creates a new empty uncle block.

pub fn into_header(self) -> H

Consumes the block and returns the header.

pub fn into_body(self) -> BlockBody<T, H>

Consumes the block and returns the body.

pub fn map_header<U>(self, f: impl FnMut(H) -> U) -> Block<T, U>

Converts the block’s header type by applying a function to it.

pub fn try_map_header<U, E>( self, f: impl FnMut(H) -> Result<U, E>, ) -> Result<Block<T, U>, E>

Converts the block’s header type by applying a fallible function to it.

pub fn convert_transactions<U>(self) -> Block<U, H>
where U: From<T>,

Converts the block’s transaction type to the given alternative that is From<T>

pub fn try_convert_transactions<U>( self, ) -> Result<Block<U, H>, <U as TryFrom<T>>::Error>
where U: TryFrom<T>,

Converts the block’s transaction to the given alternative that is TryFrom<T>

Returns the block with the new transaction type if all conversions were successful.

pub fn map_transactions<U>(self, f: impl FnMut(T) -> U) -> Block<U, H>

Converts the block’s transaction type by applying a function to each transaction.

Returns the block with the new transaction type.

pub fn try_map_transactions<U, E>( self, f: impl FnMut(T) -> Result<U, E>, ) -> Result<Block<U, H>, E>

Converts the block’s transaction type by applying a fallible function to each transaction.

Returns the block with the new transaction type if all transactions were successfully.

pub fn into_with_encoded2718(self) -> Block<WithEncoded<T>, H>
where T: Encodable2718,

Converts the transactions in the block’s body to WithEncoded<T> by encoding them via Encodable2718

pub fn with_header(self, header: H) -> Block<T, H>

Replaces the header of the block.

Note: This method only replaces the main block header. If you need to transform the ommer headers as well, use map_header instead.

pub fn rlp_encoded_from_parts(header: &H, body: &BlockBody<T, H>) -> Vec<u8>
where H: Encodable, T: Encodable,

Encodes the Block given header and block body.

Returns the rlp encoded block.

This is equivalent to block.encode.

pub fn rlp_encode_from_parts( header: &H, body: &BlockBody<T, H>, out: &mut dyn BufMut, )
where H: Encodable, T: Encodable,

Encodes the Block given header and block body

This is equivalent to block.encode.

pub fn rlp_length_for(header: &H, body: &BlockBody<T, H>) -> usize
where H: Encodable, T: Encodable,

Returns the RLP encoded length of the block’s header and body.

impl<T, H> Block<T, H>
where T: Decodable, H: Decodable,

pub fn decode_sealed(buf: &mut &[u8]) -> Result<Sealed<Block<T, H>>, Error>

Decodes the block from RLP, computing the header hash directly from the RLP bytes.

This is more efficient than decoding the block and then sealing it, as the header hash is computed from the raw RLP bytes without re-encoding.