hopr_chain_types::exports::alloy::consensus::transaction

Struct TxEip4844WithSidecar

pub struct TxEip4844WithSidecar<T = BlobTransactionSidecarVariant> {
    pub tx: TxEip4844,
    pub sidecar: T,
}

EIP-4844 Blob Transaction

A transaction with blob hashes and max blob fee, which also includes the BlobTransactionSidecar. This is the full type sent over the network as a raw transaction. It wraps a TxEip4844 to include the sidecar and the ability to decode it properly.

This is defined in EIP-4844 as an element of a PooledTransactions response, and is also used as the format for sending raw transactions through the network (eth_sendRawTransaction/eth_sendTransaction).

Fields
tx: TxEip4844

The actual transaction.

sidecar: T

The sidecar.

Implementations

impl<T> TxEip4844WithSidecar<T>

pub const fn from_tx_and_sidecar( tx: TxEip4844, sidecar: T, ) -> TxEip4844WithSidecar<T>

Constructs a new TxEip4844WithSidecar from a TxEip4844 and a sidecar.

pub const fn tx_type() -> TxType

Get the transaction type.

pub const fn tx(&self) -> &TxEip4844

Get access to the inner tx TxEip4844.

pub const fn sidecar(&self) -> &T

Get access to the inner sidecar.

pub fn into_sidecar(self) -> T

Consumes the TxEip4844WithSidecar and returns the inner sidecar.

pub fn into_parts(self) -> (TxEip4844, T)

Consumes the TxEip4844WithSidecar and returns the inner TxEip4844 and a sidecar.

pub fn map_sidecar<U>(self, f: impl FnOnce(T) -> U) -> TxEip4844WithSidecar<U>

Maps the sidecar to a new type.

pub fn try_map_sidecar<U, E>( self, f: impl FnOnce(T) -> Result<U, E>, ) -> Result<TxEip4844WithSidecar<U>, E>

Maps the sidecar to a new type, returning an error if the mapping fails.

impl<T> TxEip4844WithSidecar<T>
where T: TxEip4844Sidecar,

pub fn size(&self) -> usize

Calculates a heuristic for the in-memory size of the TxEip4844WithSidecar transaction.