hopr_chain_types::exports::alloy::consensus::transaction

Trait TxHashRef

pub trait TxHashRef {
    // Required method
    fn tx_hash(&self) -> &FixedBytes<32>;
}

Trait for types that provide access to a transaction hash reference.

This trait is implemented by types that contain or can provide a reference to a transaction hash (TxHash). It provides a standard interface for accessing transaction hashes without requiring ownership.

Required Methods

fn tx_hash(&self) -> &FixedBytes<32>

Returns a reference to the transaction hash.

This assumes the implementing type already owns or has computed the transaction hash.