hopr_chain_types::exports::alloy::contract

Struct Event

pub struct Event<P, E, N = Ethereum> {
    pub provider: P,
    pub filter: Filter,
    /* private fields */
}

Helper for managing the event filter before querying or streaming its logs

Fields
provider: P

The provider to use for querying or streaming logs.

filter: Filter

The filter to use for querying or streaming logs.

Implementations

impl<P, E, N> Event<P, E, N>
where P: Provider<N>, E: SolEvent, N: Network,

pub const fn new(provider: P, filter: Filter) -> Event<P, E, N>

Creates a new event with the provided provider and filter.

pub async fn query(&self) -> Result<Vec<(E, Log)>, Error>

Queries the blockchain for the selected filter and returns a vector of matching event logs.

pub async fn query_raw(&self) -> Result<Vec<Log>, RpcError<TransportErrorKind>>

Queries the blockchain for the selected filter and returns a vector of matching event logs, without decoding them.

pub async fn watch( &self, ) -> Result<EventPoller<E>, RpcError<TransportErrorKind>>

Watches for events that match the filter.

Returns a stream of decoded events and raw logs.

pub fn select(self, filter: impl Into<FilterBlockOption>) -> Event<P, E, N>

Sets the inner filter object

See Filter::select.

pub fn from_block<B>(self, block: B) -> Event<P, E, N>
where B: Into<BlockNumberOrTag>,

Sets the from block number

pub fn to_block<B>(self, block: B) -> Event<P, E, N>
where B: Into<BlockNumberOrTag>,

Sets the to block number

pub fn is_pending_block_filter(&self) -> bool

Return true if filter configured to match pending block.

This means that both from_block and to_block are set to the pending tag.

pub fn at_block_hash<A>(self, hash: A) -> Event<P, E, N>
where A: Into<FixedBytes<32>>,

Pins the block hash for the filter

pub fn address<A>(self, address: A) -> Event<P, E, N>
where A: Into<ValueOrArray<Address>>,

Sets the address to query with this filter.

See Filter::address.

pub fn event(self, event_name: &str) -> Event<P, E, N>

Given the event signature in string form, it hashes it and adds it to the topics to monitor

pub fn events( self, events: impl IntoIterator<Item = impl AsRef<[u8]>>, ) -> Event<P, E, N>

Hashes all event signatures and sets them as array to event_signature(topic0)

pub fn event_signature<TO>(self, topic: TO) -> Event<P, E, N>
where TO: Into<FilterSet<FixedBytes<32>>>,

Sets event_signature(topic0) (the event name for non-anonymous events)

pub fn topic1<TO>(self, topic: TO) -> Event<P, E, N>
where TO: Into<FilterSet<FixedBytes<32>>>,

Sets the 1st indexed topic

pub fn topic2<TO>(self, topic: TO) -> Event<P, E, N>
where TO: Into<FilterSet<FixedBytes<32>>>,

Sets the 2nd indexed topic

pub fn topic3<TO>(self, topic: TO) -> Event<P, E, N>
where TO: Into<FilterSet<FixedBytes<32>>>,

Sets the 3rd indexed topic

impl<P, E, N> Event<&P, E, N>
where P: Clone,

pub fn with_cloned_provider(self) -> Event<P, E, N>

Clones the provider and returns a new event with the cloned provider.