hopr_chain_types::exports::alloy::core::json_abi

Struct JsonAbi

pub struct JsonAbi {
    pub constructor: Option<Constructor>,
    pub fallback: Option<Fallback>,
    pub receive: Option<Receive>,
    pub functions: BTreeMap<String, Vec<Function>>,
    pub events: BTreeMap<String, Vec<Event>>,
    pub errors: BTreeMap<String, Vec<Error>>,
}

The JSON contract ABI, as specified in the Solidity ABI spec.

Fields
constructor: Option<Constructor>

The constructor function.

fallback: Option<Fallback>

The fallback function.

receive: Option<Receive>

The receive function.

functions: BTreeMap<String, Vec<Function>>

The functions, indexed by the function name.

events: BTreeMap<String, Vec<Event>>

The events, indexed by the event name.

errors: BTreeMap<String, Vec<Error>>

The errors, indexed by the error name.

Implementations

impl JsonAbi

pub fn new() -> JsonAbi

Creates an empty ABI object.

pub fn parse<'a, I>(strings: I) -> Result<JsonAbi, Error>
where I: IntoIterator<Item = &'a str>,

Parse a Human-Readable ABI string into a JSON object.

Examples

assert_eq!(JsonAbi::parse([])?, JsonAbi::new());

let abi = JsonAbi::parse([
    "constructor(string symbol, string name)",
    "function transferFrom(address from, address to, uint value)",
    "function balanceOf(address owner)(uint balance)",
    "event Transfer(address indexed from, address indexed to, address value)",
    "error InsufficientBalance(address owner, uint balance)",
    "function addPerson(tuple(string, uint16) person)",
    "function addPeople(tuple(string, uint16)[] person)",
    "function getPerson(uint id)(tuple(string, uint16))",
    "event PersonAdded(uint indexed id, tuple(string, uint16) person)",
])?;
assert_eq!(abi.len(), 9);

pub fn len(&self) -> usize

Returns the total number of items (of any type).

pub fn is_empty(&self) -> bool

Returns true if the ABI contains no items.

pub fn items(&self) -> Items<'_>

Returns an iterator over all of the items in the ABI.

pub fn into_items(self) -> IntoItems

Returns an iterator over all of the items in the ABI.

pub fn to_sol(&self, name: &str, config: Option<ToSolConfig>) -> String

Formats this JSON ABI as a Solidity interface.

The order of the definitions is not guaranteed.

Generates:

interface <name> {
    <enums>...
    <UDVTs>...
    <structs>...
    <errors>...
    <events>...
    <fallback>
    <receive>
    <functions>...
}

Note that enums are going to be identical to uint8 UDVTs, since no other information about enums is present in the ABI.

pub fn to_sol_raw( &self, name: &str, out: &mut String, config: Option<ToSolConfig>, )

Formats this JSON ABI as a Solidity interface into the given string.

See to_sol for more information.

pub fn dedup(&mut self)

Deduplicates all functions, errors, and events which have the same name and inputs.

pub const fn constructor(&self) -> Option<&Constructor>

Returns an immutable reference to the constructor.

pub fn constructor_mut(&mut self) -> Option<&mut Constructor>

Returns a mutable reference to the constructor.

pub fn function(&self, name: &str) -> Option<&Vec<Function>>

Returns an immutable reference to the list of all the functions with the given name.

pub fn function_mut(&mut self, name: &str) -> Option<&mut Vec<Function>>

Returns a mutable reference to the list of all the functions with the given name.

pub fn event(&self, name: &str) -> Option<&Vec<Event>>

Returns an immutable reference to the list of all the events with the given name.

pub fn event_mut(&mut self, name: &str) -> Option<&mut Vec<Event>>

Returns a mutable reference to the list of all the events with the given name.

pub fn error(&self, name: &str) -> Option<&Vec<Error>>

Returns an immutable reference to the list of all the errors with the given name.

pub fn error_mut(&mut self, name: &str) -> Option<&mut Vec<Error>>

Returns a mutable reference to the list of all the errors with the given name.

pub fn functions(&self) -> Flatten<Values<'_, String, Vec<Function>>>

Returns an iterator over immutable references to the functions.

pub fn functions_mut(&mut self) -> Flatten<ValuesMut<'_, String, Vec<Function>>>

Returns an iterator over mutable references to the functions.

pub fn function_by_selector(&self, selector: FixedBytes<4>) -> Option<&Function>

Returns the first Function with a matching selector.

pub fn events(&self) -> Flatten<Values<'_, String, Vec<Event>>>

Returns an iterator over immutable references to the events.

pub fn events_mut(&mut self) -> Flatten<ValuesMut<'_, String, Vec<Event>>>

Returns an iterator over mutable references to the events.

pub fn errors(&self) -> Flatten<Values<'_, String, Vec<Error>>>

Returns an iterator over immutable references to the errors.

pub fn errors_mut(&mut self) -> Flatten<ValuesMut<'_, String, Vec<Error>>>

Returns an iterator over mutable references to the errors.