hopr_chain_types::exports::alloy::core::primitives::ruint::algorithms::div

Function div_nx1

pub unsafe fn div_nx1(limbs: &mut [u64], divisor: u64) -> u64

⚠️ Compute single limb division.

⚠️ This function is not part of the stable API.
See algorithm 7 from [MG10].
Safety

The following conditions must hold:

The highest limb of numerator must be nonzero.

The highest limb of divisor must be nonzero.

In debug, may panic if any condition of use is violated. In release, may panic or trigger UB if any condition of use is violated.