hopr_chain_types::exports::alloy::core::primitives

Struct LogData

pub struct LogData {
    pub data: Bytes,
    /* private fields */
}

An Ethereum event log object.

Fields
data: Bytes

The plain data.

Implementations

impl LogData

pub const fn new_unchecked(topics: Vec<FixedBytes<32>>, data: Bytes) -> LogData

Creates a new log, without length-checking. This allows creation of invalid logs. May be safely used when the length of the topic list is known to be 4 or less.

pub fn new(topics: Vec<FixedBytes<32>>, data: Bytes) -> Option<LogData>

Creates a new log.

pub const fn empty() -> LogData

Creates a new empty log.

pub fn is_valid(&self) -> bool

True if valid, false otherwise.

pub fn topics(&self) -> &[FixedBytes<32>]

Get the topic list.

pub fn topics_mut(&mut self) -> &mut [FixedBytes<32>]

Get the topic list, mutably. This gives access to the internal array, without allowing extension of that array.

pub const fn topics_mut_unchecked(&mut self) -> &mut Vec<FixedBytes<32>>

Get a mutable reference to the topic list. This allows creation of invalid logs.

pub fn set_topics_unchecked(&mut self, topics: Vec<FixedBytes<32>>)

Set the topic list, without length-checking. This allows creation of invalid logs.

pub fn set_topics_truncating(&mut self, topics: Vec<FixedBytes<32>>)

Set the topic list, truncating to 4 topics.

pub fn split(self) -> (Vec<FixedBytes<32>>, Bytes)

Consumes the log data, returning the topic list and the data.