hopr_chain_types::exports::alloy::core::sol_types

Trait SolError

pub trait SolError: Sized {
    type Parameters<'a>: SolType<Token<'a> = Self::Token<'a>>;
    type Token<'a>: TokenSeq<'a>;

    const SIGNATURE: &'static str;
    const SELECTOR: [u8; 4];

    // Required methods
    fn new(tuple: <Self::Parameters<'_> as SolType>::RustType) -> Self;
    fn tokenize(&self) -> Self::Token<'_>;

    // Provided methods
    fn abi_encoded_size(&self) -> usize { ... }
    fn abi_decode_raw(data: &[u8]) -> Result<Self, Error> { ... }
    fn abi_decode_raw_validate(data: &[u8]) -> Result<Self, Error> { ... }
    fn abi_decode(data: &[u8]) -> Result<Self, Error> { ... }
    fn abi_decode_validate(data: &[u8]) -> Result<Self, Error> { ... }
    fn abi_encode_raw(&self, out: &mut Vec<u8>) { ... }
    fn abi_encode(&self) -> Vec<u8>  { ... }
}

A Solidity custom error.

Implementer’s Guide

It should not be necessary to implement this trait manually. Instead, use the sol! procedural macro to parse Solidity syntax into types that implement this trait.

Required Associated Constants

const SIGNATURE: &'static str

The error’s ABI signature.

const SELECTOR: [u8; 4]

The error selector: keccak256(SIGNATURE)[0..4]

Required Associated Types

type Parameters<'a>: SolType<Token<'a> = Self::Token<'a>>

The underlying tuple type which represents the error’s members.

If the error has no arguments, this will be the unit type ()

type Token<'a>: TokenSeq<'a>

The corresponding TokenSeq type.

Required Methods

fn new(tuple: <Self::Parameters<'_> as SolType>::RustType) -> Self

Convert from the tuple type used for ABI encoding and decoding.

fn tokenize(&self) -> Self::Token<'_>

Convert to the token type used for EIP-712 encoding and decoding.

Provided Methods

fn abi_encoded_size(&self) -> usize

The size of the error params when encoded in bytes, without the selector.

fn abi_decode_raw(data: &[u8]) -> Result<Self, Error>

ABI decode this call’s arguments from the given slice, without its selector.

fn abi_decode_raw_validate(data: &[u8]) -> Result<Self, Error>

ABI decode this error’s arguments from the given slice, without its selector, with validation.

This is the same as abi_decode_raw, but performs validation checks on the decoded parameters tuple.

fn abi_decode(data: &[u8]) -> Result<Self, Error>

ABI decode this error’s arguments from the given slice, with the selector.

fn abi_decode_validate(data: &[u8]) -> Result<Self, Error>

ABI decode this error’s arguments from the given slice, with the selector, with validation.

This is the same as abi_decode, but performs validation checks on the decoded parameters tuple.

fn abi_encode_raw(&self, out: &mut Vec<u8>)

ABI encode the error to the given buffer without its selector.

fn abi_encode(&self) -> Vec<u8>

ABI encode the error to the given buffer with its selector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.