hopr_chain_types::exports::alloy::node_bindings

Struct Anvil

pub struct Anvil { /* private fields */ }

Builder for launching anvil.

Panics

If spawn is called without anvil being available in the userโs $PATH

Example

use alloy_node_bindings::Anvil;

let port = 8545u16;
let url = format!("http://localhost:{}", port).to_string();

let anvil = Anvil::new()
    .port(port)
    .mnemonic("abstract vacuum mammal awkward pudding scene penalty purchase dinner depart evoke puzzle")
    .spawn();

drop(anvil); // this will kill the instance

Implementations

impl Anvil

pub fn new() -> Anvil

Creates an empty Anvil builder. The default port and the mnemonic are chosen randomly.

Example

fn a() {
 let anvil = Anvil::default().spawn();

 println!("Anvil running at `{}`", anvil.endpoint());

pub fn at(path: impl Into<PathBuf>) -> Anvil

Creates an Anvil builder which will execute anvil at the given path.

Example

fn a() {
 let anvil = Anvil::at("~/.foundry/bin/anvil").spawn();

 println!("Anvil running at `{}`", anvil.endpoint());

pub fn path<T>(self, path: T) -> Anvil
where T: Into<PathBuf>,

Sets the path to the anvil cli

By default, itโs expected that anvil is in $PATH, see also std::process::Command::new()

pub fn host<T>(self, host: T) -> Anvil
where T: Into<String>,

Sets the host which will be used when the anvil instance is launched.

pub fn port<T>(self, port: T) -> Anvil
where T: Into<u16>,

Sets the port which will be used when the anvil instance is launched.

pub fn ipc_path(self, path: impl Into<String>) -> Anvil

Sets the path for the ipc server

pub const fn chain_id(self, chain_id: u64) -> Anvil

Sets the chain_id the anvil instance will use.

If not set, the instance defaults to chain id 31337.

pub fn mnemonic<T>(self, mnemonic: T) -> Anvil
where T: Into<String>,

Sets the mnemonic which will be used when the anvil instance is launched.

pub const fn block_time(self, block_time: u64) -> Anvil

Sets the block-time in seconds which will be used when the anvil instance is launched.

pub const fn block_time_f64(self, block_time: f64) -> Anvil

Sets the block-time in sub-seconds which will be used when the anvil instance is launched. Older versions of anvil do not support sub-second block times.

pub const fn fork_block_number(self, fork_block_number: u64) -> Anvil

Sets the fork-block-number which will be used in addition to Self::fork.

Note: if set, then this requires fork to be set as well

pub fn fork<T>(self, fork: T) -> Anvil
where T: Into<String>,

Sets the fork argument to fork from another currently running Ethereum client at a given block. Input should be the HTTP location and port of the other client, e.g. http://localhost:8545. You can optionally specify the block to fork from using an @ sign: http://localhost:8545@1599200

pub fn hardfork(self, hardfork: EthereumHardfork) -> Anvil

Select the [EthereumHardfork] to start anvil with.

pub fn paris(self) -> Anvil

Set the [EthereumHardfork] to [EthereumHardfork::Paris].

pub fn cancun(self) -> Anvil

Set the [EthereumHardfork] to [EthereumHardfork::Cancun].

pub fn shanghai(self) -> Anvil

Set the [EthereumHardfork] to [EthereumHardfork::Shanghai].

pub fn prague(self) -> Anvil

Set the [EthereumHardfork] to [EthereumHardfork::Prague].

pub fn odyssey(self) -> Anvil

Instantiate anvil with the --odyssey flag.

pub fn auto_impersonate(self) -> Anvil

Instantiate anvil with the --auto-impersonate flag.

pub fn push_arg<T>(&mut self, arg: T)
where T: Into<OsString>,

Adds an argument to pass to the anvil.

pub fn extend_args<I, S>(&mut self, args: I)
where I: IntoIterator<Item = S>, S: Into<OsString>,

Adds multiple arguments to pass to the anvil.

pub fn arg<T>(self, arg: T) -> Anvil
where T: Into<OsString>,

Adds an argument to pass to the anvil.

pub fn args<I, S>(self, args: I) -> Anvil
where I: IntoIterator<Item = S>, S: Into<OsString>,

Adds multiple arguments to pass to the anvil.

pub fn env<K, V>(self, key: K, value: V) -> Anvil
where K: Into<OsString>, V: Into<OsString>,

Adds an environment variable to pass to the anvil.

pub fn envs<I, K, V>(self, envs: I) -> Anvil
where I: IntoIterator<Item = (K, V)>, K: Into<OsString>, V: Into<OsString>,

Adds multiple environment variables to pass to the anvil.

pub const fn timeout(self, timeout: u64) -> Anvil

Sets the timeout which will be used when the anvil instance is launched. Units: milliseconds.

pub const fn keep_stdout(self) -> Anvil

Keep the handle to anvilโs stdout in order to read from it.

Caution: if the stdout handle isnโt used, this can end up blocking.

pub fn spawn(self) -> AnvilInstance

Consumes the builder and spawns anvil.

Panics

If spawning the instance fails at any point.

pub fn try_spawn(self) -> Result<AnvilInstance, NodeError>

Consumes the builder and spawns anvil. If spawning fails, returns an error.