hopr_chain_types::exports::alloy::providers

Enum ProviderCall

pub enum ProviderCall<Params, Resp, Output = Resp, Map = fn(Resp) -> Output>
where
    Params: RpcSend,
    Resp: RpcRecv,
    Map: Fn(Resp) -> Output,
{
    RpcCall(RpcCall<Params, Resp, Output, Map>),
    Waiter(Waiter<Resp, Output, Map>),
    BoxedFuture(Pin<Box<dyn Future<Output = Result<Output, RpcError<TransportErrorKind>>> + Send>>),
    Ready(Option<Result<Output, RpcError<TransportErrorKind>>>),
}

The primary future type for the Provider.

This future abstracts over several potential data sources. It allows providers to:

produce data via an RpcCall

produce data by waiting on a batched RPC Waiter

proudce data via an arbitrary boxed future

produce data in any synchronous way

Variants

RpcCall(RpcCall<Params, Resp, Output, Map>)

An underlying call to an RPC server.

Waiter(Waiter<Resp, Output, Map>)

A waiter for a batched call to a remote RPC server.

BoxedFuture(Pin<Box<dyn Future<Output = Result<Output, RpcError<TransportErrorKind>>> + Send>>)

A boxed future.

Ready(Option<Result<Output, RpcError<TransportErrorKind>>>)

The output, produces synchronously.

Implementations

impl<Params, Resp, Output, Map> ProviderCall<Params, Resp, Output, Map>
where Params: RpcSend, Resp: RpcRecv, Map: Fn(Resp) -> Output,

pub const fn ready( output: Result<Output, RpcError<TransportErrorKind>>, ) -> ProviderCall<Params, Resp, Output, Map>

Instantiate a new ProviderCall from the output.

pub const fn is_rpc_call(&self) -> bool

True if this is an RPC call.

pub const fn as_rpc_call(&self) -> Option<&RpcCall<Params, Resp, Output, Map>>

Fallible cast to RpcCall

pub const fn as_mut_rpc_call( &mut self, ) -> Option<&mut RpcCall<Params, Resp, Output, Map>>

Fallible cast to mutable RpcCall

pub const fn is_waiter(&self) -> bool

True if this is a waiter.

pub const fn as_waiter(&self) -> Option<&Waiter<Resp, Output, Map>>

Fallible cast to Waiter

pub const fn as_mut_waiter(&mut self) -> Option<&mut Waiter<Resp, Output, Map>>

Fallible cast to mutable Waiter

pub const fn is_boxed_future(&self) -> bool

True if this is a boxed future.

pub const fn as_boxed_future( &self, ) -> Option<&Pin<Box<dyn Future<Output = Result<Output, RpcError<TransportErrorKind>>> + Send>>>

Fallible cast to a boxed future.

pub const fn is_ready(&self) -> bool

True if this is a ready value.

pub const fn as_ready( &self, ) -> Option<&Result<Output, RpcError<TransportErrorKind>>>

Fallible cast to a ready value.

Panics

Panics if the future is already complete

pub fn map_resp<NewOutput, NewMap>( self, map: NewMap, ) -> Result<ProviderCall<Params, Resp, NewOutput, NewMap>, ProviderCall<Params, Resp, Output, Map>>
where NewMap: Fn(Resp) -> NewOutput + Clone,

Set a function to map the response into a different type. This is useful for transforming the response into a more usable type, e.g. changing U64 to u64.

This function fails if the inner future is not an RpcCall or Waiter.

Note

Carefully review the rust documentation on fn pointers before passing them to this function. Unless the pointer is specifically coerced to a fn(_) -> _, the NewMap will be inferred as that function’s unique type. This can lead to confusing error messages.

impl<Params, Resp, Output, Map> ProviderCall<&Params, Resp, Output, Map>
where Params: RpcSend + ToOwned, <Params as ToOwned>::Owned: RpcSend, Resp: RpcRecv, Map: Fn(Resp) -> Output,

pub fn into_owned_params( self, ) -> ProviderCall<<Params as ToOwned>::Owned, Resp, Output, Map>

Convert this call into one with owned params, by cloning the params.

Panics

Panics if called after the request has been polled.