hopr_chain_types::exports::alloy::providers

Module fillers

Transaction fillers.

Fillers decorate a Provider, filling transaction details before they are sent to the network, like nonces, gas limits, and gas prices.

Fillers are called before any other layer in the provider.

Implementing a filler

Fillers implement the TxFiller trait. Before a filler is called, TxFiller::status is called to determine whether the filler has any work to do. If this function returns FillerControlFlow::Ready, the filler will be called.

Composing fillers

To layer fillers, a utility filler is provided called JoinFill, which is a composition of two fillers, left and right. The left filler is called before the right filler.

Structs

BlobGasFiller

Filler for the max_fee_per_blob_gas field in EIP-4844 transactions.

CachedNonceManager

Cached nonce manager

ChainIdFiller

A TxFiller that populates the chain ID of a transaction.

FillProvider

A Provider that applies one or more TxFillers.

GasFiller

A TxFiller that populates gas related fields in transaction requests if unset.

JoinFill

A filler that can fill in a TransactionRequest with additional information by joining two TxFillers.

NonceFiller

A TxFiller that fills nonces on transactions. The behavior of filling nonces is determined by the NonceManager.

SimpleNonceManager

This NonceManager implementation will fetch the transaction count for any new account it sees.

WalletFiller

A layer that signs transactions locally.

Enums

BlobGasEstimator

Blob Gas estimator variants

FillEnvelopeError

Error type for failures in the fill_envelope function.

FillerControlFlow

The control flow for a filler.

Traits

BlobGasEstimatorFn

A trait responsible for estimating blob gas values

NonceManager

A trait that determines the behavior of filling nonces.

RecommendedFillers

A trait which may be used to configure default fillers for Network implementations.

TxFiller

A layer that can fill in a TransactionRequest with additional information.

Type Aliases

BlobGasEstimatorFunction

An estimator function for blob gas fees.

RecommendedFiller

The recommended filler, a preconfigured set of layers handling gas estimation, nonce management, and chain-id fetching.