hopr_chain_types::exports::alloy::providers::network::primitives

Trait TransactionResponse

pub trait TransactionResponse: Transaction {
    // Required methods
    fn tx_hash(&self) -> FixedBytes<32>;
    fn block_hash(&self) -> Option<FixedBytes<32>>;
    fn block_number(&self) -> Option<u64>;
    fn transaction_index(&self) -> Option<u64>;
    fn from(&self) -> Address;

    // Provided methods
    fn block_hash_num(&self) -> Option<NumHash> { ... }
    fn gas_price(&self) -> Option<u128> { ... }
    fn max_fee_per_gas(&self) -> Option<u128> { ... }
    fn transaction_type(&self) -> Option<u8> { ... }
    fn inclusion_info(&self) -> Option<InclusionInfo> { ... }
}

Transaction JSON-RPC response. Aggregates transaction data with its block and signer context.

Required Methods

fn tx_hash(&self) -> FixedBytes<32>

Hash of the transaction

fn block_hash(&self) -> Option<FixedBytes<32>>

Returns the hash of the block this transaction was mined in.

Returns None if this transaction is still pending.

fn block_number(&self) -> Option<u64>

Returns the number of the block this transaction was mined in.

Returns None if this transaction is still pending.

fn transaction_index(&self) -> Option<u64>

Transaction Index

fn from(&self) -> Address

Sender of the transaction

Provided Methods

fn block_hash_num(&self) -> Option<NumHash>

Returns the BlockNumHash of the block this transaction was mined in.

Returns None if this transaction is still pending.

fn gas_price(&self) -> Option<u128>

Gas Price, this is the RPC format for max_fee_per_gas, pre-eip-1559.

fn max_fee_per_gas(&self) -> Option<u128>

Max BaseFeePerGas the user is willing to pay. For pre-eip-1559 transactions, the field label gas_price is used instead.

fn transaction_type(&self) -> Option<u8>

Transaction type format for RPC. This field is included since eip-2930.

fn inclusion_info(&self) -> Option<InclusionInfo>

Returns the InclusionInfo if the transaction has been included.

Returns None if this transaction is still pending (missing block number, hash, or index).