hopr_chain_types::exports::alloy::providers::network

Trait ReceiptResponse

pub trait ReceiptResponse {

    // Required methods
    fn contract_address(&self) -> Option<Address>;
    fn status(&self) -> bool;
    fn block_hash(&self) -> Option<FixedBytes<32>>;
    fn block_number(&self) -> Option<u64>;
    fn transaction_hash(&self) -> FixedBytes<32>;
    fn transaction_index(&self) -> Option<u64>;
    fn gas_used(&self) -> u64;
    fn effective_gas_price(&self) -> u128;
    fn blob_gas_used(&self) -> Option<u64>;
    fn blob_gas_price(&self) -> Option<u128>;
    fn from(&self) -> Address;
    fn to(&self) -> Option<Address>;
    fn cumulative_gas_used(&self) -> u64;
    fn state_root(&self) -> Option<FixedBytes<32>>;

    // Provided methods
    fn block_hash_num(&self) -> Option<NumHash> { ... }
    fn cost(&self) -> u128 { ... }
    fn ensure_success(&self) -> Result<(), TransactionFailedError> { ... }

}

Receipt JSON-RPC response.

Required Methods

fn contract_address(&self) -> Option<Address>

Address of the created contract, or None if the transaction was not a deployment.

fn status(&self) -> bool

Status of the transaction.

Note

Caution must be taken when using this method for deep-historical receipts, as it may not accurately reflect the status of the transaction. The transaction status is not knowable from the receipt for transactions before [EIP-658].

fn block_hash(&self) -> Option<FixedBytes<32>>

Hash of the block this transaction was included within.

fn block_number(&self) -> Option<u64>

Number of the block this transaction was included within.

fn transaction_hash(&self) -> FixedBytes<32>

Transaction Hash.

fn transaction_index(&self) -> Option<u64>

Index within the block.

fn gas_used(&self) -> u64

Gas used by this transaction alone.

fn effective_gas_price(&self) -> u128

Effective gas price.

fn blob_gas_used(&self) -> Option<u64>

Blob gas used by the eip-4844 transaction.

fn blob_gas_price(&self) -> Option<u128>

Blob gas price paid by the eip-4844 transaction.

fn from(&self) -> Address

Address of the sender.

fn to(&self) -> Option<Address>

Address of the receiver.

fn cumulative_gas_used(&self) -> u64

Returns the cumulative gas used at this receipt.

fn state_root(&self) -> Option<FixedBytes<32>>

The post-transaction state root (pre Byzantium)

EIP98 makes this field optional.

Provided Methods

fn block_hash_num(&self) -> Option<NumHash>

Returns the BlockNumHash of the block this transaction was mined in.

Returns None if this transaction is still pending.

fn cost(&self) -> u128

Total cost of this transaction = gas_used * effective_gas_price.

fn ensure_success(&self) -> Result<(), TransactionFailedError>

Ensures the transaction was successful, returning an error if it failed.