hopr_chain_types::exports::alloy::providers

Struct ProviderBuilder

pub struct ProviderBuilder<L, F, N = Ethereum> { /* private fields */ }

A builder for constructing a Provider from various layers.

This type is similar to tower::ServiceBuilder, with extra complication around maintaining the network and transport types.

The ProviderBuilder can be instantiated in two ways, using ProviderBuilder::new() or ProviderBuilder::default().

ProviderBuilder::new() will create a new ProviderBuilder with the RecommendedFillers enabled, whereas ProviderBuilder::default() will instantiate it in its vanilla ProviderBuilder form i.e with no fillers enabled.

Implementations

impl ProviderBuilder<Identity, JoinFill<Identity, <Ethereum as RecommendedFillers>::RecommendedFillers>>

pub fn new() -> ProviderBuilder<Identity, JoinFill<Identity, <Ethereum as RecommendedFillers>::RecommendedFillers>>

Create a new ProviderBuilder with the recommended filler enabled.

Recommended fillers are preconfigured set of fillers that handle gas estimation, nonce management, and chain-id fetching.

Building a provider with this setting enabled will return a crate::fillers::FillProvider with crate::utils::JoinedRecommendedFillers.

You can opt-out of using these fillers by using the .disable_recommended_fillers() method.

pub fn disable_recommended_fillers(self) -> ProviderBuilder<Identity, Identity>

Opt-out of the recommended fillers by resetting the fillers stack in the ProviderBuilder.

This is equivalent to creating the builder using ProviderBuilder::default().

impl ProviderBuilder<Identity, Identity>

pub fn new_with_network<Net>() -> ProviderBuilder<Identity, JoinFill<Identity, <Net as RecommendedFillers>::RecommendedFillers>, Net>
where Net: RecommendedFillers,

Create a new ProviderBuilder with the RecommendedFillers for the provided Network.

impl<L, N> ProviderBuilder<L, Identity, N>
where N: Network,

pub fn with_recommended_fillers( self, ) -> ProviderBuilder<L, JoinFill<Identity, <N as RecommendedFillers>::RecommendedFillers>, N>
where N: RecommendedFillers,

Add preconfigured set of layers handling gas estimation, nonce management, and chain-id fetching.

impl<L, F, N> ProviderBuilder<L, F, N>

pub fn layer<Inner>( self, layer: Inner, ) -> ProviderBuilder<Stack<Inner, L>, F, N>

Add a layer to the stack being built. This is similar to tower::ServiceBuilder::layer.

Note:

Layers are added in outer-to-inner order, as in tower::ServiceBuilder. The first layer added will be the first to see the request.

pub fn filler<F2>(self, filler: F2) -> ProviderBuilder<L, JoinFill<F, F2>, N>

Add a transaction filler to the stack being built. Transaction fillers are used to fill in missing fields on transactions before they are sent, and are all joined to form the outermost layer of the stack.

pub fn network<Net>(self) -> ProviderBuilder<L, F, Net>
where Net: Network,

Change the network.

By default, the network is Ethereum. This method must be called to configure a different network.

builder.network::<Arbitrum>()

pub fn with_chain( self, chain: NamedChain, ) -> ProviderBuilder<Stack<ChainLayer, L>, F, N>

Add a chain layer to the stack being built. The layer will set the client’s poll interval based on the average block time for this chain.

Does nothing to the client with a local transport.

pub fn with_blob_gas_estimation( self, ) -> ProviderBuilder<L, JoinFill<F, BlobGasFiller>, N>

Add blob gas estimation to the stack being built.

See BlobGasFiller for more information.

pub fn with_blob_gas_estimator( self, estimator: BlobGasEstimator, ) -> ProviderBuilder<L, JoinFill<F, BlobGasFiller>, N>

Add blob gas estimation to the stack being built, using the provided estimator.

See BlobGasFiller and BlobGasEstimator for more information.

pub fn with_gas_estimation( self, ) -> ProviderBuilder<L, JoinFill<F, GasFiller>, N>

Add gas estimation to the stack being built.

See GasFiller for more information.

pub fn with_nonce_management<M>( self, nonce_manager: M, ) -> ProviderBuilder<L, JoinFill<F, NonceFiller<M>>, N>
where M: NonceManager,

Add nonce management to the stack being built.

See NonceFiller for more information.

pub fn with_simple_nonce_management( self, ) -> ProviderBuilder<L, JoinFill<F, NonceFiller<SimpleNonceManager>>, N>

Add simple nonce management to the stack being built.

See SimpleNonceManager for more information.

pub fn with_cached_nonce_management( self, ) -> ProviderBuilder<L, JoinFill<F, NonceFiller>, N>

Add cached nonce management to the stack being built.

See CachedNonceManager for more information.

pub fn fetch_chain_id(self) -> ProviderBuilder<L, JoinFill<F, ChainIdFiller>, N>

Add a chain ID filler to the stack being built. The filler will attempt to fetch the chain ID from the provider using Provider::get_chain_id. the first time a transaction is prepared, and will cache it for future transactions.

pub fn with_chain_id( self, chain_id: u64, ) -> ProviderBuilder<L, JoinFill<F, ChainIdFiller>, N>

Add a specific chain ID to the stack being built. The filler will fill transactions with the provided chain ID, regardless of the chain ID that the provider reports via Provider::get_chain_id.

pub fn wallet<W>( self, wallet: W, ) -> ProviderBuilder<L, JoinFill<F, WalletFiller<<W as IntoWallet<N>>::NetworkWallet>>, N>
where W: IntoWallet<N>, N: Network,

Add a wallet layer to the stack being built.

See WalletFiller.

pub fn with_call_batching( self, ) -> ProviderBuilder<Stack<CallBatchLayer, L>, F, N>

Aggregate multiple eth_call requests into a single batch request using Multicall3.

See CallBatchLayer for more information.

pub fn with_arbitrum_call_batching( self, ) -> ProviderBuilder<Stack<CallBatchLayer, L>, F, N>

Aggregate multiple eth_call requests with block number queries done by calling Arbsym precompile.

See CallBatchLayer for more information.

pub fn with_caching( self, max_items: u32, ) -> ProviderBuilder<Stack<CacheLayer, L>, F, N>

Add response caching to the stack being built with the specified maximum cache size.

See CacheLayer for more information.

pub fn with_default_caching(self) -> ProviderBuilder<Stack<CacheLayer, L>, F, N>

Add response caching to the stack being built with a default cache size of 100 items.

See CacheLayer for more information.

pub fn with_default_block( self, block_id: BlockId, ) -> ProviderBuilder<Stack<BlockIdLayer, L>, F, N>

Set a default BlockId for eth_call and eth_estimateGas.

pub fn connect_provider<P>( self, provider: P, ) -> <F as ProviderLayer<<L as ProviderLayer<P, N>>::Provider, N>>::Provider
where L: ProviderLayer<P, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<P, N>>::Provider, N>, P: Provider<N>, N: Network,

Finish the layer stack by providing a root Provider, outputting the final Provider type with all stack components.

pub fn connect_client( self, client: RpcClient, ) -> <F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network,

Finish the layer stack by providing a root RpcClient, outputting the final Provider type with all stack components.

This is a convenience function for ProviderBuilder::on_provider(RootProvider::new(client)).

pub fn connect_mocked_client( self, asserter: Asserter, ) -> <F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network,

Finish the layer stack by providing a RpcClient that mocks responses, outputting the final Provider type with all stack components.

This is a convenience function for ProviderBuilder::on_client(RpcClient::mocked(asserter)).

pub async fn connect( self, s: &str, ) -> Result<<F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider, RpcError<TransportErrorKind>>
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network,

Finish the layer stack by providing a connection string for a built-in transport type, outputting the final Provider type with all stack components.

pub async fn connect_with_config( self, s: &str, config: ConnectionConfig, ) -> Result<<F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider, RpcError<TransportErrorKind>>
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network,

Finish the layer stack by providing a connection string with custom configuration.

This method allows for fine-grained control over connection settings such as authentication, retry behavior, and transport-specific options. The transport type is extracted from the connection string and configured using the provided ConnectionConfig.

Examples

use alloy_provider::{ConnectionConfig, ProviderBuilder};
use alloy_transport::Authorization;
use std::time::Duration;

let config = ConnectionConfig::new()
    .with_auth(Authorization::bearer("my-token"))
    .with_max_retries(3)
    .with_retry_interval(Duration::from_secs(2));

let provider =
    ProviderBuilder::new().connect_with_config("ws://localhost:8545", config).await?;

pub async fn connect_with<C>( self, connect: &C, ) -> Result<<F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider, RpcError<TransportErrorKind>>
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network, C: TransportConnect,

Finish the layer stack by providing a TransportConnect instance.

pub fn connect_http( self, url: Url, ) -> <F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network,

Build this provider with an Reqwest HTTP transport.

pub fn connect_reqwest<C>( self, client: C, url: Url, ) -> <F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network, C: Into<Client>,

Build this provider with a pre-built Reqwest client.

pub fn with_reqwest<B>( self, url: Url, builder: B, ) -> <F as ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<RootProvider<N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<RootProvider<N>, N>>::Provider, N>, N: Network, B: FnOnce(ClientBuilder) -> Client,

Build this provider with a provided Reqwest client builder.

impl<L, F, N> ProviderBuilder<L, F, N>
where N: Network,

pub fn connect_anvil( self, ) -> <F as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider
where F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>, L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>,

Build this provider with anvil, using the BoxTransport.

pub fn connect_anvil_with_wallet( self, ) -> <JoinFill<F, WalletFiller<EthereumWallet>> as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider
where F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>, L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>, EthereumWallet: NetworkWallet<N>,

Build this provider with anvil, using the BoxTransport. This function configures a wallet backed by anvil keys, and is intended for use in tests.

pub fn connect_anvil_with_config( self, f: impl FnOnce(Anvil) -> Anvil, ) -> <F as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider
where F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>, L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>,

Build this provider with anvil, using the BoxTransport. The given function is used to configure the anvil instance.

pub fn on_anvil_with_config( self, f: impl FnOnce(Anvil) -> Anvil, ) -> <F as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider
where L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>, F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>,

👎Deprecated since 0.12.6: use connect_anvil_with_config instead

Build this provider with anvil, using the BoxTransport. The given function is used to configure the anvil instance.

pub fn connect_anvil_with_wallet_and_config( self, f: impl FnOnce(Anvil) -> Anvil, ) -> Result<<JoinFill<F, WalletFiller<EthereumWallet>> as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider, NodeError>
where F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>, L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>, EthereumWallet: NetworkWallet<N>,

Build this provider with anvil, using the BoxTransport. This calls try_on_anvil_with_wallet_and_config and panics on error.

pub fn on_anvil_with_wallet_and_config( self, f: impl FnOnce(Anvil) -> Anvil, ) -> Result<<JoinFill<F, WalletFiller<EthereumWallet>> as ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>>::Provider, NodeError>
where F: TxFiller<N> + ProviderLayer<<L as ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>>::Provider, N>, L: ProviderLayer<AnvilProvider<RootProvider<N>, N>, N>, EthereumWallet: NetworkWallet<N>,

👎Deprecated since 0.12.6: use connect_anvil_with_wallet_and_config instead

Build this provider with anvil, using the BoxTransport. This calls try_on_anvil_with_wallet_and_config and panics on error.