hopr_chain_types::exports::alloy::rpc::client

Struct RpcClient

pub struct RpcClient(/* private fields */);

A JSON-RPC client.

RpcClient should never be instantiated directly. Instead, use ClientBuilder.

Implementations

impl RpcClient

pub const fn builder() -> ClientBuilder<Identity>

Create a new ClientBuilder.

impl RpcClient

pub fn new(t: impl IntoBoxTransport, is_local: bool) -> RpcClient

Creates a new RpcClient with the given transport.

pub fn mocked(asserter: Asserter) -> RpcClient

Create a new RpcClient with a transport that returns mocked responses from the given Asserter.

pub fn new_http(url: Url) -> RpcClient

Create a new RpcClient with an HTTP transport.

pub fn new_http_with_client(client: Client, url: Url) -> RpcClient

Create a new RpcClient with an HTTP transport using a pre-built reqwest::Client.

pub fn from_inner(inner: RpcClientInner) -> RpcClient

Creates a new RpcClient with the given inner client.

pub const fn inner(&self) -> &Arc<RpcClientInner>

Get a reference to the client.

pub fn into_inner(self) -> Arc<RpcClientInner>

Convert the client into its inner type.

pub fn get_weak(&self) -> Weak<RpcClientInner>

Get a Weak reference to the client.

pub fn get_ref(&self) -> &RpcClientInner

Borrow the client.

pub fn with_poll_interval(self, poll_interval: Duration) -> RpcClient

Sets the poll interval for the client in milliseconds.

Note: RpcClient internally wraps an Arc<RpcClientInner>, so updating the poll interval changes it for all RpcClient handles that share the same inner client.

pub fn prepare_static_poller<Params, Resp>( &self, method: impl Into<Cow<'static, str>>, params: Params, ) -> PollerBuilder<Params, Resp>
where Params: RpcSend + 'static, Resp: RpcRecv + Clone,

Build a poller that polls a method with the given parameters.

See PollerBuilder for examples and more details.

pub fn new_batch(&self) -> BatchRequest<'_>

Create a new BatchRequest builder.