hopr_chain_types::exports::alloy::rpc::json_rpc

Struct Response

pub struct Response<Payload = Box<RawValue>, ErrData = Box<RawValue>> {
    pub id: Id,
    pub payload: ResponsePayload<Payload, ErrData>,
}

A JSON-RPC 2.0 response object containing a ResponsePayload.

This object is used to represent a JSON-RPC 2.0 response. It may contain either a successful result or an error. The id field is used to match the response to the request that it is responding to, and should be mirrored from the response.

Fields
id: Id

The ID of the request that this response is responding to.

payload: ResponsePayload<Payload, ErrData>

The response payload.

Implementations

impl Response<&RawValue, &RawValue>

pub fn into_owned(self) -> Response

Convert this borrowed response to an owned response by copying the data from the deserializer (if necessary).

impl<Payload, ErrData> Response<Payload, ErrData>

pub const fn parse_error(id: Id) -> Response<Payload, ErrData>

Create a new response with a parsed error payload.

pub const fn invalid_request(id: Id) -> Response<Payload, ErrData>

Create a new response with an invalid request error payload.

pub const fn method_not_found(id: Id) -> Response<Payload, ErrData>

Create a new response with a method not found error payload.

pub const fn invalid_params(id: Id) -> Response<Payload, ErrData>

Create a new response with an invalid params error payload.

pub const fn internal_error(id: Id) -> Response<Payload, ErrData>

Create a new response with an internal error payload.

pub const fn internal_error_message( id: Id, message: Cow<'static, str>, ) -> Response<Payload, ErrData>

Create a new error response for an internal error with a custom message.

pub const fn payload(&self) -> &ResponsePayload<Payload, ErrData>

Returns the payload of this response

pub const fn internal_error_with_obj( id: Id, data: ErrData, ) -> Response<Payload, ErrData>
where ErrData: RpcSend,

Create a new error response for an internal error with additional data.

pub const fn internal_error_with_message_and_obj( id: Id, message: Cow<'static, str>, data: ErrData, ) -> Response<Payload, ErrData>
where ErrData: RpcSend,

Create a new error response for an internal error with a custom message and additional data.

pub const fn is_success(&self) -> bool

Returns true if the response is a success.

pub const fn is_error(&self) -> bool

Returns true if the response is an error.

pub fn error_code(&self) -> Option<i64>

Returns the error code if the payload of this response is an ErrorPayload.

impl<Payload, ErrData> Response<Payload, ErrData>
where Payload: RpcSend, ErrData: RpcSend,

pub fn serialize_payload(&self) -> Result<Response, Error>

Serialize the payload of this response.

impl<'a, Payload, ErrData> Response<Payload, ErrData>
where Payload: AsRef<RawValue> + 'a,

pub fn try_success_as<T>(&'a self) -> Option<Result<T, Error>>
where T: Deserialize<'a>,

Attempt to deserialize the success payload, borrowing from the payload if necessary.

See ResponsePayload::try_success_as.

pub fn deser_success<T>( self, ) -> Result<Response<T, ErrData>, Response<Payload, ErrData>>
where T: DeserializeOwned,

Attempt to deserialize the Success payload, transforming this type.

Returns

Ok(Response<T, ErrData>) if the payload is a success and can be deserialized as T, or if the payload is an error.

Err(self) if the payload is a success and can’t be deserialized.

impl<'a, Payload, ErrData> Response<Payload, ErrData>
where ErrData: Borrow<RawValue> + 'a,

pub fn try_error_as<T>(&'a self) -> Option<Result<T, Error>>
where T: Deserialize<'a>,

Attempt to deserialize the error payload, borrowing from the payload if necessary.

See ResponsePayload::try_error_as.

pub fn deser_err<T>( self, ) -> Result<Response<Payload, T>, Response<Payload, ErrData>>
where T: DeserializeOwned,

Attempt to deserialize the Error payload, transforming this type.

Returns

Ok(Response<Payload, T>) if the payload is an error and can be deserialized as T, or if the payload is a success.

Err(self) if the payload is an error and can’t be deserialized.