hopr_chain_types::exports::alloy::rpc::json_rpc

Type Alias BorrowedResponse

pub type BorrowedResponse<'a> = Response<&'a RawValue, &'a RawValue>;

A Response that has been partially deserialized, borrowing its contents from the deserializer. This is used primarily for intermediate deserialization. Most users will not require it.

See the top-level docs for more info.

Aliased Type

pub struct BorrowedResponse<'a> {
    pub id: Id,
    pub payload: ResponsePayload<&'a RawValue, &'a RawValue>,
}

Fields
id: Id

The ID of the request that this response is responding to.

payload: ResponsePayload<&'a RawValue, &'a RawValue>

The response payload.