hopr_chain_types::exports::alloy::rpc::types::eth

Struct AnyReceiptEnvelope

pub struct AnyReceiptEnvelope<T = Log> {
    pub inner: ReceiptWithBloom<Receipt<T>>,
    pub type: u8,
}

Receipt envelope, as defined in EIP-2718.

This enum distinguishes between tagged and untagged legacy receipts, as the in-protocol Merkle tree may commit to EITHER 0-prefixed or raw. Therefore we must ensure that encoding returns the precise byte-array that was decoded, preserving the presence or absence of the TransactionType flag.

Transaction receipt payloads are specified in their respective EIPs.

Fields
inner: ReceiptWithBloom<Receipt<T>>

The receipt envelope.

type: u8

The transaction type.

Implementations

impl<T> AnyReceiptEnvelope<T>

pub const fn is_legacy(&self) -> bool

Returns whether this is a legacy receipt (type 0)

impl<T> AnyReceiptEnvelope<T>
where T: Encodable,

pub fn rlp_payload_length(&self) -> usize

Calculate the length of the rlp payload of the network encoded receipt.

impl<T> AnyReceiptEnvelope<T>

pub const fn is_success(&self) -> bool

Return true if the transaction was successful.

Note

This method may not accurately reflect the status of the transaction for transactions before EIP-658.

pub const fn status(&self) -> bool

Returns the success status of the receipt’s transaction.

Note

This method may not accurately reflect the status of the transaction for transactions before EIP-658.

pub const fn bloom(&self) -> Bloom

Return the receipt’s bloom.

pub const fn bloom_ref(&self) -> &Bloom

Return a reference to the receipt’s bloom.

pub const fn cumulative_gas_used(&self) -> u64

Returns the cumulative gas used at this receipt.

pub fn logs(&self) -> &[T]

Return the receipt logs.