hopr_chain_types::exports::alloy::rpc::types::eth

Struct Transaction

pub struct Transaction<T = EthereumTxEnvelope<TxEip4844Variant>> {
    pub inner: Recovered<T>,
    pub block_hash: Option<FixedBytes<32>>,
    pub block_number: Option<u64>,
    pub transaction_index: Option<u64>,
    pub effective_gas_price: Option<u128>,
}

Transaction object used in RPC.

This represents a transaction in RPC format (eth_getTransactionByHash) and contains the full transaction object and additional block metadata if the transaction has been mined.

Fields
inner: Recovered<T>

The inner transaction object

block_hash: Option<FixedBytes<32>>

Hash of block where transaction was included, None if pending

block_number: Option<u64>

Number of block where transaction was included, None if pending

transaction_index: Option<u64>

Transaction Index

effective_gas_price: Option<u128>

Deprecated effective gas price value.

Implementations

impl<T> Transaction<T>

pub fn into_inner(self) -> T

Consumes the type and returns the wrapped transaction.

pub fn into_recovered(self) -> Recovered<T>

Consumes the type and returns a Recovered transaction with the sender

pub const fn as_recovered(&self) -> Recovered<&T>

Returns a Recovered<&T> with the transaction and the sender.

pub fn convert<U>(self) -> Transaction<U>
where U: From<T>,

Converts the transaction type to the given alternative that is From<T>

pub fn try_convert<U>(self) -> Result<Transaction<U>, <U as TryFrom<T>>::Error>
where U: TryFrom<T>,

Converts the transaction to the given alternative that is TryFrom<T>

Returns the transaction with the new transaction type if all conversions were successful.

pub fn map<Tx>(self, f: impl FnOnce(T) -> Tx) -> Transaction<Tx>

Applies the given closure to the inner transaction type.

pub fn try_map<Tx, E>( self, f: impl FnOnce(T) -> Result<Tx, E>, ) -> Result<Transaction<Tx>, E>

Applies the given fallible closure to the inner transactions.

impl<T> Transaction<T>
where T: Transaction,

pub fn is_legacy_gas(&self) -> bool

Returns true if the transaction is a legacy or 2930 transaction.

pub fn from_transaction( tx: Recovered<T>, tx_info: TransactionInfo, ) -> Transaction<T>

Converts a consensus tx with an additional context tx_info into an RPC Transaction.

impl<T> Transaction<T>
where T: Transaction + Encodable2718,

pub fn info(&self) -> TransactionInfo

Returns the TransactionInfo for this transaction.

This contains various metadata about the transaction and block context if available.

impl<T> Transaction<T>
where T: Into<TransactionRequest>,

pub fn into_request(self) -> TransactionRequest

Converts Transaction into TransactionRequest.

During this conversion data for TransactionRequest::sidecar is not populated as it is not part of Transaction.

impl<Eip4844> Transaction<EthereumTxEnvelope<Eip4844>>

pub fn into_signed(self) -> Signed<EthereumTypedTransaction<Eip4844>>
where EthereumTypedTransaction<Eip4844>: From<Eip4844>,

Consumes the transaction and returns it as Signed with EthereumTypedTransaction as the transaction type.

pub fn into_signed_recovered( self, ) -> Recovered<Signed<EthereumTypedTransaction<Eip4844>>>
where EthereumTypedTransaction<Eip4844>: From<Eip4844>,

Consumes the transaction and returns it a Recovered signed EthereumTypedTransaction.