hopr_chain_types::exports::alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum

Trait ToInt

pub trait ToInt<T> {
    const INT: T;

    // Required method
    fn to_int() -> T;
}

A type operator for taking a concrete integer value from a type.

It returns arbitrary integer value without explicitly specifying the type. It is useful when you pass the values to methods that accept distinct types without runtime casting.

Required Associated Constants

const INT: T

The concrete value for the type. Can be used in const contexts.

Required Methods

fn to_int() -> T

Method returning the concrete value for the type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.