hopr_chain_types::exports::alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum

Type Alias U

pub type U<const N: usize> = <Const<N> as ToUInt>::Output;

The main mapping from a generic const: usize to a UInt: U<N> is expected to work like UN.

It requires the const-generics crate feature to be enabled.

Example

use typenum::*;

assert_type_eq!(U<42>, U42);

This can even be used in a generic const N: usize context, provided the genericity is guarded by a where clause:

use typenum::*;

struct MyStruct<const N: usize>;

trait MyTrait { type AssocType; }

impl<const N: usize> MyTrait
    for MyStruct<N>
where
    Const<N> : ToUInt,
{
    type AssocType = U<N>;
}

assert_type_eq!(<MyStruct<42> as MyTrait>::AssocType, U42);