hopr_chain_types::exports::alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum::type_operators

Trait Max

pub trait Max<Rhs = Self> {
    type Output;

    // Required method
    fn max(self, rhs: Rhs) -> Self::Output;
}

A type operator that returns the maximum of Self and Rhs.

Required Associated Types

type Output

The type of the maximum of Self and Rhs

Required Methods

fn max(self, rhs: Rhs) -> Self::Output

Method returning the maximum