hopr_chain_types::exports::alloy::signers::k256::ecdsa::signature

Trait Verifier

pub trait Verifier<S> {
    // Required method
    fn verify(&self, msg: &[u8], signature: &S) -> Result<(), Error>;
}

Verify the provided message bytestring using Self (e.g. a public key)

Required Methods

fn verify(&self, msg: &[u8], signature: &S) -> Result<(), Error>

Use Self to verify that the provided signature for a given message bytestring is authentic.

Returns Error if it is inauthentic, or otherwise returns ().