hopr_chain_types::exports::alloy::signers::k256::elliptic_curve::bigint::zeroize

Trait Zeroize

pub trait Zeroize {
    // Required method
    fn zeroize(&mut self);
}

Trait for securely erasing values from memory.

Required Methods

fn zeroize(&mut self)

Zero out this object from memory using Rust intrinsics which ensure the zeroization operation is not “optimized away” by the compiler.