hopr_chain_types::exports::alloy::signers::k256::elliptic_curve

Crate ff

This crate provides traits for working with finite fields.

Modules

helpers

Helper methods for implementing the ff traits.

Structs

BatchInverter

A non-allocating batch inverter.

Traits

BatchInvert

Extension trait for iterators over mutable field elements which allows those field elements to be inverted in a batch.

Field

This trait represents an element of a field.

FromUniformBytes

Trait for constructing a PrimeField element from a fixed-length uniform byte array.

PrimeField

This represents an element of a non-binary prime field.

WithSmallOrderMulGroup

The subset of prime-order fields such that (modulus - 1) is divisible by N.