hopr_chain_types::exports::alloy::signers::k256::elliptic_curve

Module hash2curve

Traits for hashing byte sequences to curve points.

https://datatracker.ietf.org/doc/draft-irtf-cfrg-hash-to-curve

Structs

ExpandMsgXmd

Placeholder type for implementing expand_message_xmd based on a hash function

ExpandMsgXof

Placeholder type for implementing expand_message_xof based on an extendable output function

ExpanderXmd

Expander type for ExpandMsgXmd.

IsogenyCoefficients

The coefficients for mapping from one isogenous curve to another

OsswuMapParams

The Optimized Simplified Shallue-van de Woestijne-Ulas parameters

Traits

ExpandMsg

Trait for types implementing expand_message interface for hash_to_field.

Expander

Expander that, call read until enough bytes have been consumed.

FromOkm

The trait for helping to convert to a field element.

GroupDigest

Adds hashing arbitrary byte sequences to a valid group element

Isogeny

The Isogeny methods to map to another curve.

MapToCurve

Trait for converting field elements into a point via a mapping method like Simplified Shallue-van de Woestijne-Ulas or Elligator

OsswuMap

The optimized simplified Shallue-van de Woestijne-Ulas method for mapping elliptic curve scalars to affine points.

Sgn0

Trait for determining the parity of the field