hopr_chain_types::exports::alloy::signers::k256::elliptic_curve::pkcs8::der::referenced

Trait OwnedToRef

pub trait OwnedToRef {
    type Borrowed<'a>
       where Self: 'a;

    // Required method
    fn owned_to_ref(&self) -> Self::Borrowed<'_>;
}

A trait for borrowing data from an owned struct

Required Associated Types

type Borrowed<'a> where Self: 'a

The resulting type referencing back to Self

Required Methods

fn owned_to_ref(&self) -> Self::Borrowed<'_>

Creates a new object referencing back to the self for storage

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.