hopr_chain_types::exports::alloy::signers::k256::elliptic_curve

Struct PublicKey

pub struct PublicKey<C>
where
    C: CurveArithmetic,
{ /* private fields */ }

Elliptic curve public keys.

This is a wrapper type for AffinePoint which ensures an inner non-identity point and provides a common place to handle encoding/decoding.

Parsing “SPKI” Keys

X.509 SubjectPublicKeyInfo (SPKI) is a commonly used format for encoding public keys, notably public keys corresponding to PKCS#8 private keys. (especially ones generated by OpenSSL).

Keys in SPKI format are either binary (ASN.1 BER/DER), or PEM encoded (ASCII) and begin with the following:

-----BEGIN PUBLIC KEY-----

To decode an elliptic curve public key from SPKI, enable the pkcs8 feature of this crate (or the pkcs8 feature of a specific RustCrypto elliptic curve crate) and use the elliptic_curve::pkcs8::DecodePublicKey trait to parse it.

When the pem feature of this crate (or a specific RustCrypto elliptic curve crate) is enabled, a [FromStr] impl is also available.

serde support

When the optional serde feature of this create is enabled, Serialize and Deserialize impls are provided for this type.

The serialization is binary-oriented and supports ASN.1 DER Subject Public Key Info (SPKI) as the encoding format.

For a more text-friendly encoding of public keys, use [JwkEcKey] instead.

Implementations

impl<C> PublicKey<C>
where C: CurveArithmetic,

pub fn from_affine( point: <C as CurveArithmetic>::AffinePoint, ) -> Result<PublicKey<C>, Error>

Convert an AffinePoint into a PublicKey

pub fn from_secret_scalar(scalar: &NonZeroScalar<C>) -> PublicKey<C>

Compute a PublicKey from a secret NonZeroScalar value (i.e. a secret key represented as a raw scalar value)

pub fn from_sec1_bytes(bytes: &[u8]) -> Result<PublicKey<C>, Error>
where <C as Curve>::FieldBytesSize: ModulusSize, <C as CurveArithmetic>::AffinePoint: FromEncodedPoint<C> + ToEncodedPoint<C>,

Decode PublicKey (compressed or uncompressed) from the Elliptic-Curve-Point-to-Octet-String encoding described in SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.3 (page 10).

http://www.secg.org/sec1-v2.pdf

pub fn to_sec1_bytes(&self) -> Box<[u8]>
where C: PointCompression, <C as CurveArithmetic>::AffinePoint: FromEncodedPoint<C> + ToEncodedPoint<C>, <C as Curve>::FieldBytesSize: ModulusSize,

Convert this PublicKey into the Elliptic-Curve-Point-to-Octet-String encoding described in SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.3 (page 10).

http://www.secg.org/sec1-v2.pdf

pub fn as_affine(&self) -> &<C as CurveArithmetic>::AffinePoint

Borrow the inner AffinePoint from this PublicKey.

In ECC, public keys are elliptic curve points.

pub fn to_projective(&self) -> <C as CurveArithmetic>::ProjectivePoint

Convert this PublicKey to a ProjectivePoint for the given curve

pub fn to_nonidentity(&self) -> NonIdentity<<C as CurveArithmetic>::AffinePoint>

Convert this PublicKey to a NonIdentity of the inner AffinePoint