hopr_chain_types::exports::alloy::signers::local

Struct LocalSigner

pub struct LocalSigner<C> { /* private fields */ }

An Ethereum private-public key pair which can be used for signing messages.

Examples

Signing and Verifying a message

The signer can be used to produce ECDSA Signature objects, which can be then verified. Note that this uses eip191_hash_message under the hood which will prefix the message being hashed with the Ethereum Signed Message domain separator.

use alloy_signer::{Signer, SignerSync};
use alloy_signer_local::PrivateKeySigner;

let signer = PrivateKeySigner::random();

// Optionally, the signer's chain id can be set, in order to use EIP-155
// replay protection with different chains
let signer = signer.with_chain_id(Some(1337));

// The signer can be used to sign messages
let message = b"hello";
let signature = signer.sign_message_sync(message)?;
assert_eq!(signature.recover_address_from_msg(&message[..]).unwrap(), signer.address());

// LocalSigner is cloneable:
let signer_clone = signer.clone();
let signature2 = signer_clone.sign_message_sync(message)?;
assert_eq!(signature, signature2);

Implementations

impl LocalSigner<SigningKey<Secp256k1>>

pub fn from_signing_key( credential: SigningKey<Secp256k1>, ) -> LocalSigner<SigningKey<Secp256k1>>

Creates a new LocalSigner instance from a SigningKey.

This can also be used to create a LocalSigner from a SecretKey. See also the From implementations.

pub fn from_bytes( bytes: &FixedBytes<32>, ) -> Result<LocalSigner<SigningKey<Secp256k1>>, Error>

Creates a new LocalSigner instance from a raw scalar serialized as a B256 byte array.

This is identical to from_field_bytes.

pub fn from_field_bytes( bytes: &GenericArray<u8, <Secp256k1 as Curve>::FieldBytesSize>, ) -> Result<LocalSigner<SigningKey<Secp256k1>>, Error>

Creates a new LocalSigner instance from a raw scalar serialized as a FieldBytes byte array.

pub fn from_slice( bytes: &[u8], ) -> Result<LocalSigner<SigningKey<Secp256k1>>, Error>

Creates a new LocalSigner instance from a raw scalar serialized as a byte slice.

Byte slices shorter than the field size (32 bytes) are handled by zero padding the input.

pub fn random() -> LocalSigner<SigningKey<Secp256k1>>

Creates a new random keypair seeded with rand::thread_rng().

pub fn random_with<R>(rng: &mut R) -> LocalSigner<SigningKey<Secp256k1>>
where R: Rng + CryptoRng,

Creates a new random keypair seeded with the provided RNG.

pub fn as_nonzero_scalar(&self) -> &NonZeroScalar<Secp256k1>

Borrow the secret NonZeroScalar value for this key.

⚠️ Warning

This value is key material.

Please treat it with the care it deserves!

pub fn to_bytes(&self) -> FixedBytes<32>

Serialize this LocalSigner’s SigningKey as a B256 byte array.

pub fn to_field_bytes( &self, ) -> GenericArray<u8, <Secp256k1 as Curve>::FieldBytesSize>

Serialize this LocalSigner’s SigningKey as a FieldBytes byte array.

pub fn public_key(&self) -> FixedBytes<64>

Convenience function that returns this signer’s ethereum public key as a B512 byte array.

impl<C> LocalSigner<C>
where C: PrehashSigner<(Signature<Secp256k1>, RecoveryId)>,

pub const fn new_with_credential( credential: C, address: Address, chain_id: Option<u64>, ) -> LocalSigner<C>

Construct a new credential with an external PrehashSigner.

pub const fn credential(&self) -> &C

Returns this signer’s credential.

pub fn into_credential(self) -> C

Consumes this signer and returns its credential.

pub const fn address(&self) -> Address

Returns this signer’s address.

pub const fn chain_id(&self) -> Option<u64>

Returns this signer’s chain ID.