hopr_chain_types::exports::alloy::transports::http::reqwest::dns

Trait Resolve

pub trait Resolve: Send + Sync {
    // Required method
    fn resolve(
        &self,
        name: Name,
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Iterator<Item = SocketAddr> + Send>, Box<dyn Error + Sync + Send>>> + Send>>;
}

Trait for customizing DNS resolution in reqwest.

Required Methods

fn resolve( &self, name: Name, ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Iterator<Item = SocketAddr> + Send>, Box<dyn Error + Sync + Send>>> + Send>>

Performs DNS resolution on a Name. The return type is a future containing an iterator of SocketAddr.

It differs from tower_service::Service<Name> in several ways:

It is assumed that resolve will always be ready to poll.

It does not need a mutable reference to self.

Since trait objects cannot make use of associated types, it requires wrapping the returned Future and its contained Iterator with Box.

Explicitly specified port in the URL will override any port in the resolved SocketAddrs. Otherwise, port 0 will be replaced by the conventional port for the given scheme (e.g. 80 for http).