hopr_chain_types::exports::alloy::transports::http::reqwest

Struct ClientBuilder

pub struct ClientBuilder { /* private fields */ }

A ClientBuilder can be used to create a Client with custom configuration.

Implementations

impl ClientBuilder

pub fn new() -> ClientBuilder

Constructs a new ClientBuilder.

This is the same as Client::builder().

impl ClientBuilder

pub fn build(self) -> Result<Client, Error>

Returns a Client that uses this ClientBuilder configuration.

Errors

This method fails if a TLS backend cannot be initialized, or the resolver cannot load the system configuration.

pub fn user_agent<V>(self, value: V) -> ClientBuilder
where V: TryInto<HeaderValue>, <V as TryInto<HeaderValue>>::Error: Into<Error>,

Sets the User-Agent header to be used by this client.

Example

// Name your user agent after your app?
static APP_USER_AGENT: &str = concat!(
    env!("CARGO_PKG_NAME"),
    "/",
    env!("CARGO_PKG_VERSION"),
);

let client = reqwest::Client::builder()
    .user_agent(APP_USER_AGENT)
    .build()?;
let res = client.get("https://www.rust-lang.org").send().await?;

pub fn default_headers(self, headers: HeaderMap) -> ClientBuilder

Sets the default headers for every request.

Example

use reqwest::header;
let mut headers = header::HeaderMap::new();
headers.insert("X-MY-HEADER", header::HeaderValue::from_static("value"));

// Consider marking security-sensitive headers with `set_sensitive`.
let mut auth_value = header::HeaderValue::from_static("secret");
auth_value.set_sensitive(true);
headers.insert(header::AUTHORIZATION, auth_value);

// get a client builder
let client = reqwest::Client::builder()
    .default_headers(headers)
    .build()?;
let res = client.get("https://www.rust-lang.org").send().await?;

pub fn gzip(self, enable: bool) -> ClientBuilder

Enable auto gzip decompression by checking the Content-Encoding response header.

If auto gzip decompression is turned on:

When sending a request and if the request’s headers do not already contain an Accept-Encoding and Range values, the Accept-Encoding header is set to gzip. The request body is not automatically compressed.

When receiving a response, if its headers contain a Content-Encoding value of gzip, both Content-Encoding and Content-Length are removed from the headers’ set. The response body is automatically decompressed.

If the gzip feature is turned on, the default option is enabled.

Optional

This requires the optional gzip feature to be enabled

pub fn brotli(self, enable: bool) -> ClientBuilder

Enable auto brotli decompression by checking the Content-Encoding response header.

If auto brotli decompression is turned on:

When sending a request and if the request’s headers do not already contain an Accept-Encoding and Range values, the Accept-Encoding header is set to br. The request body is not automatically compressed.

When receiving a response, if its headers contain a Content-Encoding value of br, both Content-Encoding and Content-Length are removed from the headers’ set. The response body is automatically decompressed.

If the brotli feature is turned on, the default option is enabled.

Optional

This requires the optional brotli feature to be enabled

pub fn zstd(self, enable: bool) -> ClientBuilder

Enable auto zstd decompression by checking the Content-Encoding response header.

If auto zstd decompression is turned on:

When sending a request and if the request’s headers do not already contain an Accept-Encoding and Range values, the Accept-Encoding header is set to zstd. The request body is not automatically compressed.

When receiving a response, if its headers contain a Content-Encoding value of zstd, both Content-Encoding and Content-Length are removed from the headers’ set. The response body is automatically decompressed.

If the zstd feature is turned on, the default option is enabled.

Optional

This requires the optional zstd feature to be enabled

pub fn deflate(self, enable: bool) -> ClientBuilder

Enable auto deflate decompression by checking the Content-Encoding response header.

If auto deflate decompression is turned on:

When sending a request and if the request’s headers do not already contain an Accept-Encoding and Range values, the Accept-Encoding header is set to deflate. The request body is not automatically compressed.

When receiving a response, if it’s headers contain a Content-Encoding value that equals to deflate, both values Content-Encoding and Content-Length are removed from the headers’ set. The response body is automatically decompressed.

If the deflate feature is turned on, the default option is enabled.

Optional

This requires the optional deflate feature to be enabled

pub fn no_gzip(self) -> ClientBuilder

Disable auto response body gzip decompression.

This method exists even if the optional gzip feature is not enabled. This can be used to ensure a Client doesn’t use gzip decompression even if another dependency were to enable the optional gzip feature.

pub fn no_brotli(self) -> ClientBuilder

Disable auto response body brotli decompression.

This method exists even if the optional brotli feature is not enabled. This can be used to ensure a Client doesn’t use brotli decompression even if another dependency were to enable the optional brotli feature.

pub fn no_zstd(self) -> ClientBuilder

Disable auto response body zstd decompression.

This method exists even if the optional zstd feature is not enabled. This can be used to ensure a Client doesn’t use zstd decompression even if another dependency were to enable the optional zstd feature.

pub fn no_deflate(self) -> ClientBuilder

Disable auto response body deflate decompression.

This method exists even if the optional deflate feature is not enabled. This can be used to ensure a Client doesn’t use deflate decompression even if another dependency were to enable the optional deflate feature.

pub fn redirect(self, policy: Policy) -> ClientBuilder

Set a RedirectPolicy for this client.

Default will follow redirects up to a maximum of 10.

pub fn referer(self, enable: bool) -> ClientBuilder

Enable or disable automatic setting of the Referer header.

Default is true.

pub fn retry(self, policy: Builder) -> ClientBuilder

Set a request retry policy.

Default behavior is to retry protocol NACKs.

pub fn proxy(self, proxy: Proxy) -> ClientBuilder

Add a Proxy to the list of proxies the Client will use.

Note

Adding a proxy will disable the automatic usage of the “system” proxy.

pub fn no_proxy(self) -> ClientBuilder

Clear all Proxies, so Client will use no proxy anymore.

Note

To add a proxy exclusion list, use crate::proxy::Proxy::no_proxy() on all desired proxies instead.

This also disables the automatic usage of the “system” proxy.

pub fn timeout(self, timeout: Duration) -> ClientBuilder

Enables a total request timeout.

The timeout is applied from when the request starts connecting until the response body has finished. Also considered a total deadline.

Default is no timeout.

pub fn read_timeout(self, timeout: Duration) -> ClientBuilder

Enables a read timeout.

The timeout applies to each read operation, and resets after a successful read. This is more appropriate for detecting stalled connections when the size isn’t known beforehand.

Default is no timeout.

pub fn connect_timeout(self, timeout: Duration) -> ClientBuilder

Set a timeout for only the connect phase of a Client.

Default is None.

Note

This requires the futures be executed in a tokio runtime with a tokio timer enabled.

pub fn connection_verbose(self, verbose: bool) -> ClientBuilder

Set whether connections should emit verbose logs.

Enabling this option will emit log messages at the TRACE level for read and write operations on connections.

pub fn pool_idle_timeout<D>(self, val: D) -> ClientBuilder
where D: Into<Option<Duration>>,

Set an optional timeout for idle sockets being kept-alive.

Pass None to disable timeout.

Default is 90 seconds.

pub fn pool_max_idle_per_host(self, max: usize) -> ClientBuilder

Sets the maximum idle connection per host allowed in the pool.

Default is usize::MAX (no limit).

pub fn http1_title_case_headers(self) -> ClientBuilder

Send headers as title case instead of lowercase.

pub fn http1_allow_obsolete_multiline_headers_in_responses( self, value: bool, ) -> ClientBuilder

Set whether HTTP/1 connections will accept obsolete line folding for header values.

Newline codepoints (\r and \n) will be transformed to spaces when parsing.

pub fn http1_ignore_invalid_headers_in_responses( self, value: bool, ) -> ClientBuilder

Sets whether invalid header lines should be silently ignored in HTTP/1 responses.

pub fn http1_allow_spaces_after_header_name_in_responses( self, value: bool, ) -> ClientBuilder

Set whether HTTP/1 connections will accept spaces between header names and the colon that follow them in responses.

Newline codepoints (\r and \n) will be transformed to spaces when parsing.

pub fn http1_only(self) -> ClientBuilder

Only use HTTP/1.

pub fn http09_responses(self) -> ClientBuilder

Allow HTTP/0.9 responses

pub fn http2_prior_knowledge(self) -> ClientBuilder

Only use HTTP/2.

pub fn http2_initial_stream_window_size( self, sz: impl Into<Option<u32>>, ) -> ClientBuilder

Sets the SETTINGS_INITIAL_WINDOW_SIZE option for HTTP2 stream-level flow control.

Default is currently 65,535 but may change internally to optimize for common uses.

pub fn http2_initial_connection_window_size( self, sz: impl Into<Option<u32>>, ) -> ClientBuilder

Sets the max connection-level flow control for HTTP2

Default is currently 65,535 but may change internally to optimize for common uses.

pub fn http2_adaptive_window(self, enabled: bool) -> ClientBuilder

Sets whether to use an adaptive flow control.

Enabling this will override the limits set in http2_initial_stream_window_size and http2_initial_connection_window_size.

pub fn http2_max_frame_size(self, sz: impl Into<Option<u32>>) -> ClientBuilder

Sets the maximum frame size to use for HTTP2.

Default is currently 16,384 but may change internally to optimize for common uses.

pub fn http2_max_header_list_size( self, max_header_size_bytes: u32, ) -> ClientBuilder

Sets the maximum size of received header frames for HTTP2.

Default is currently 16KB, but can change.

pub fn http2_keep_alive_interval( self, interval: impl Into<Option<Duration>>, ) -> ClientBuilder

Sets an interval for HTTP2 Ping frames should be sent to keep a connection alive.

Pass None to disable HTTP2 keep-alive. Default is currently disabled.

pub fn http2_keep_alive_timeout(self, timeout: Duration) -> ClientBuilder

Sets a timeout for receiving an acknowledgement of the keep-alive ping.

If the ping is not acknowledged within the timeout, the connection will be closed. Does nothing if http2_keep_alive_interval is disabled. Default is currently disabled.

pub fn http2_keep_alive_while_idle(self, enabled: bool) -> ClientBuilder

Sets whether HTTP2 keep-alive should apply while the connection is idle.

If disabled, keep-alive pings are only sent while there are open request/responses streams. If enabled, pings are also sent when no streams are active. Does nothing if http2_keep_alive_interval is disabled. Default is false.

pub fn tcp_nodelay(self, enabled: bool) -> ClientBuilder

Set whether sockets have TCP_NODELAY enabled.

Default is true.

pub fn local_address<T>(self, addr: T) -> ClientBuilder
where T: Into<Option<IpAddr>>,

Bind to a local IP Address.

Example

use std::net::IpAddr;
let local_addr = IpAddr::from([12, 4, 1, 8]);
let client = reqwest::Client::builder()
    .local_address(local_addr)
    .build()?;

pub fn interface(self, interface: &str) -> ClientBuilder

Bind connections only on the specified network interface.

This option is only available on the following operating systems:

Android

Fuchsia

Linux,

macOS and macOS-like systems (iOS, tvOS, watchOS and visionOS)

Solaris and illumos

On Android, Linux, and Fuchsia, this uses the SO_BINDTODEVICE socket option. On macOS and macOS-like systems, Solaris, and illumos, this instead uses the IP_BOUND_IF and IPV6_BOUND_IF socket options (as appropriate).

Note that connections will fail if the provided interface name is not a network interface that currently exists when a connection is established.

Example

let interface = "lo";
let client = reqwest::Client::builder()
    .interface(interface)
    .build()?;

pub fn tcp_keepalive<D>(self, val: D) -> ClientBuilder
where D: Into<Option<Duration>>,

Set that all sockets have SO_KEEPALIVE set with the supplied duration.

If None, the option will not be set.

pub fn tcp_keepalive_interval<D>(self, val: D) -> ClientBuilder
where D: Into<Option<Duration>>,

Set that all sockets have SO_KEEPALIVE set with the supplied interval.

If None, the option will not be set.

pub fn tcp_keepalive_retries<C>(self, retries: C) -> ClientBuilder
where C: Into<Option<u32>>,

Set that all sockets have SO_KEEPALIVE set with the supplied retry count.

If None, the option will not be set.

pub fn tcp_user_timeout<D>(self, val: D) -> ClientBuilder
where D: Into<Option<Duration>>,

Set that all sockets have TCP_USER_TIMEOUT set with the supplied duration.

This option controls how long transmitted data may remain unacknowledged before the connection is force-closed.

If None, the option will not be set.

pub fn unix_socket(self, path: impl UnixSocketProvider) -> ClientBuilder

Set that all connections will use this Unix socket.

If a request URI uses the https scheme, TLS will still be used over the Unix socket.

Note

This option is not compatible with any of the TCP or Proxy options. Setting this will ignore all those options previously set.

Likewise, DNS resolution will not be done on the domain name.

pub fn add_root_certificate(self, cert: Certificate) -> ClientBuilder

Add a custom root certificate.

This can be used to connect to a server that has a self-signed certificate for example.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn tls_built_in_root_certs( self, tls_built_in_root_certs: bool, ) -> ClientBuilder

Controls the use of built-in/preloaded certificates during certificate validation.

Defaults to true – built-in system certs will be used.

Bulk Option

If this value is true, all enabled system certs configured with Cargo features will be loaded.

You can set this to false, and enable only a specific source with individual methods. Do that will prevent other sources from being loaded even if their feature Cargo feature is enabled.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn danger_accept_invalid_hostnames( self, accept_invalid_hostname: bool, ) -> ClientBuilder

Controls the use of hostname verification.

Defaults to false.

Warning

You should think very carefully before you use this method. If hostname verification is not used, any valid certificate for any site will be trusted for use from any other. This introduces a significant vulnerability to man-in-the-middle attacks.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn danger_accept_invalid_certs( self, accept_invalid_certs: bool, ) -> ClientBuilder

Controls the use of certificate validation.

Defaults to false.

Warning

You should think very carefully before using this method. If invalid certificates are trusted, any certificate for any site will be trusted for use. This includes expired certificates. This introduces significant vulnerabilities, and should only be used as a last resort.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn tls_sni(self, tls_sni: bool) -> ClientBuilder

Controls the use of TLS server name indication.

Defaults to true.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn min_tls_version(self, version: Version) -> ClientBuilder

Set the minimum required TLS version for connections.

By default, the TLS backend’s own default is used.

Errors

A value of tls::Version::TLS_1_3 will cause an error with the native-tls/default-tls backend. This does not mean the version isn’t supported, just that it can’t be set as a minimum due to technical limitations.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn max_tls_version(self, version: Version) -> ClientBuilder

Set the maximum allowed TLS version for connections.

By default, there’s no maximum.

Errors

A value of tls::Version::TLS_1_3 will cause an error with the native-tls/default-tls backend. This does not mean the version isn’t supported, just that it can’t be set as a maximum due to technical limitations.

Cannot set a maximum outside the protocol versions supported by rustls with the rustls-tls backend.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn tls_info(self, tls_info: bool) -> ClientBuilder

Add TLS information as TlsInfo extension to responses.

Optional

This requires the optional default-tls, native-tls, or rustls-tls(-...) feature to be enabled.

pub fn https_only(self, enabled: bool) -> ClientBuilder

Restrict the Client to be used with HTTPS only requests.

Defaults to false.

pub fn no_hickory_dns(self) -> ClientBuilder

Disables the hickory-dns async resolver.

This method exists even if the optional hickory-dns feature is not enabled. This can be used to ensure a Client doesn’t use the hickory-dns async resolver even if another dependency were to enable the optional hickory-dns feature.

pub fn resolve(self, domain: &str, addr: SocketAddr) -> ClientBuilder

Override DNS resolution for specific domains to a particular IP address.

Set the port to 0 to use the conventional port for the given scheme (e.g. 80 for http). Ports in the URL itself will always be used instead of the port in the overridden addr.

pub fn resolve_to_addrs( self, domain: &str, addrs: &[SocketAddr], ) -> ClientBuilder

Override DNS resolution for specific domains to particular IP addresses.

Set the port to 0 to use the conventional port for the given scheme (e.g. 80 for http). Ports in the URL itself will always be used instead of the port in the overridden addr.

pub fn dns_resolver<R>(self, resolver: Arc<R>) -> ClientBuilder
where R: Resolve + 'static,

Override the DNS resolver implementation.

Pass an Arc wrapping a type implementing Resolve. Overrides for specific names passed to resolve and resolve_to_addrs will still be applied on top of this resolver.

pub fn dns_resolver2<R>(self, resolver: R) -> ClientBuilder
where R: IntoResolve,

Override the DNS resolver implementation.

Overrides for specific names passed to resolve and resolve_to_addrs will still be applied on top of this resolver.

This method will replace dns_resolver in the next breaking change.

pub fn connector_layer<L>(self, layer: L) -> ClientBuilder
where L: Layer<BoxCloneSyncService<Unnameable, Conn, Box<dyn Error + Sync + Send>>> + Clone + Send + Sync + 'static, <L as Layer<BoxCloneSyncService<Unnameable, Conn, Box<dyn Error + Sync + Send>>>>::Service: Service<Unnameable, Response = Conn, Error = Box<dyn Error + Sync + Send>> + Clone + Send + Sync + 'static, <<L as Layer<BoxCloneSyncService<Unnameable, Conn, Box<dyn Error + Sync + Send>>>>::Service as Service<Unnameable>>::Future: Send + 'static,

Adds a new Tower Layer to the base connector Service which is responsible for connection establishment.

Each subsequent invocation of this function will wrap previous layers.

If configured, the connect_timeout will be the outermost layer.

Example usage:

use std::time::Duration;

let client = reqwest::Client::builder()
                     // resolved to outermost layer, meaning while we are waiting on concurrency limit
                     .connect_timeout(Duration::from_millis(200))
                     // underneath the concurrency check, so only after concurrency limit lets us through
                     .connector_layer(tower::timeout::TimeoutLayer::new(Duration::from_millis(50)))
                     .connector_layer(tower::limit::concurrency::ConcurrencyLimitLayer::new(2))
                     .build()
                     .unwrap();