hopr_chain_types

Module payload

Module defining various Ethereum transaction payload generators for the actions.

This module defines the basic PayloadGenerator trait that describes how an action is translated into a TransactionRequest that can be submitted on-chain.

There are two main implementations:

BasicPayloadGenerator which implements generation of a direct EIP1559 transaction payload. This is currently not used by a HOPR node.

SafePayloadGenerator which implements generation of a payload that embeds the transaction data into the SAFE transaction. This is currently the main mode of HOPR node operation.

These are currently based on the hopr-bindings crate.

Structs

BasicPayloadGenerator

Generates transaction payloads that do not use Safe-compliant ABI

GasEstimation

Estimated gas parameters for a transaction.

SafePayloadGenerator

Payload generator that generates Safe-compliant ABI

TransactionRequest

Represents all transaction requests to/from RPC.

Traits

PayloadGenerator

Trait for various implementations of common on-chain transaction payloads generators.

SignableTransaction

Trait for transaction payloads that can be signed and encoded to EIP2718 format.