hopr_chain_types::payload

Trait PayloadGenerator

pub trait PayloadGenerator {
    type TxRequest: SignableTransaction + Send;

    // Required methods
    fn approve(
        &self,
        spender: Address,
        amount: HoprBalance,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn transfer<C: Currency>(
        &self,
        destination: Address,
        amount: Balance<C>,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn announce(
        &self,
        announcement: AnnouncementData,
        key_binding_fee: HoprBalance,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn fund_channel(
        &self,
        dest: Address,
        amount: HoprBalance,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn close_incoming_channel(
        &self,
        source: Address,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn initiate_outgoing_channel_closure(
        &self,
        destination: Address,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn finalize_outgoing_channel_closure(
        &self,
        destination: Address,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn redeem_ticket(
        &self,
        acked_ticket: RedeemableTicket,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn register_safe_by_node(
        &self,
        safe_addr: Address,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn deregister_node_by_safe(
        &self,
    ) -> Result<Self::TxRequest, ChainTypesError>;
    fn deploy_safe(
        &self,
        balance: HoprBalance,
        admins: &[Address],
        include_node: bool,
        nonce: [u8; 32],
    ) -> Result<Self::TxRequest, ChainTypesError>;
}

Trait for various implementations of common on-chain transaction payloads generators.

Required Associated Types

type TxRequest: SignableTransaction + Send

Required Methods

fn approve( &self, spender: Address, amount: HoprBalance, ) -> Result<Self::TxRequest, ChainTypesError>

Create an ERC20 approve transaction payload. Pre-requisite to open payment channels. The spender address is typically the HOPR Channels contract address.

fn transfer<C: Currency>( &self, destination: Address, amount: Balance<C>, ) -> Result<Self::TxRequest, ChainTypesError>

Create a ERC20 transfer transaction payload

fn announce( &self, announcement: AnnouncementData, key_binding_fee: HoprBalance, ) -> Result<Self::TxRequest, ChainTypesError>

Creates the transaction payload to announce a node on-chain.

fn fund_channel( &self, dest: Address, amount: HoprBalance, ) -> Result<Self::TxRequest, ChainTypesError>

Creates the transaction payload to open a payment channel

fn close_incoming_channel( &self, source: Address, ) -> Result<Self::TxRequest, ChainTypesError>

Creates the transaction payload to immediately close an incoming payment channel

fn initiate_outgoing_channel_closure( &self, destination: Address, ) -> Result<Self::TxRequest, ChainTypesError>

Creates the transaction payload that initiates the closure of a payment channel. Once the notice period is due, the funds can be withdrawn using a finalizeChannelClosure transaction.

fn finalize_outgoing_channel_closure( &self, destination: Address, ) -> Result<Self::TxRequest, ChainTypesError>

Creates a transaction payload that withdraws funds from an outgoing payment channel. This will succeed once the closure notice period is due.

fn redeem_ticket( &self, acked_ticket: RedeemableTicket, ) -> Result<Self::TxRequest, ChainTypesError>

Used to create the payload to claim incentives for relaying a mixnet packet.

fn register_safe_by_node( &self, safe_addr: Address, ) -> Result<Self::TxRequest, ChainTypesError>

Creates a transaction payload to register a Safe instance which is used to manage the node’s funds

fn deregister_node_by_safe(&self) -> Result<Self::TxRequest, ChainTypesError>

Creates a transaction payload to remove the Safe instance. Once succeeded, the node no longer manages the funds.

fn deploy_safe( &self, balance: HoprBalance, admins: &[Address], include_node: bool, nonce: [u8; 32], ) -> Result<Self::TxRequest, ChainTypesError>

Creates a transaction payload to deploy a new Safe instance with the initial balance transferred from the signer and admins as Safe owners.

If include_node is true, the signer will be included in the module after deployment.

The resulting transaction requires that the signer owns at least the balance of wxHOPR tokens. The given nonce must be randomly generated for each deployment.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.