hopr_chain_types

Struct AlloyAddress

#[repr(transparent)]
pub struct AlloyAddress(pub FixedBytes<20>);

An Ethereum address, 20 bytes in length.

This type is separate from FixedBytes<20> and is declared with the wrap_fixed_bytes! macro. This allows us to implement address-specific functionality.

The main difference with the generic FixedBytes implementation is that Display formats the address using its EIP-55 checksum (to_checksum). Use Debug to display the raw bytes without the checksum.

Examples

Parsing and formatting:

use alloy_primitives::{address, Address};

let checksummed = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045";
let expected = address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045");
let address = Address::parse_checksummed(checksummed, None).expect("valid checksum");
assert_eq!(address, expected);

// Format the address with the checksum
assert_eq!(address.to_string(), checksummed);
assert_eq!(address.to_checksum(None), checksummed);

// Format the compressed checksummed address
assert_eq!(format!("{address:#}"), "0xd8dA…6045");

// Format the address without the checksum
assert_eq!(format!("{address:?}"), "0xd8da6bf26964af9d7eed9e03e53415d37aa96045");

Tuple Fields
0: FixedBytes<20>
Implementations

impl Address

pub const ZERO: Address

Array of Zero bytes.

pub const fn new(bytes: [u8; 20]) -> Address

Wraps the given byte array in this type.

pub const fn with_last_byte(x: u8) -> Address

Creates a new byte array with the last byte set to x.

pub const fn repeat_byte(byte: u8) -> Address

Creates a new byte array where all bytes are set to byte.

pub const fn len_bytes() -> usize

Returns the size of this array in bytes.

pub fn from_slice(src: &[u8]) -> Address

Create a new byte array from the given slice src.

For a fallible version, use the TryFrom<&[u8]> implementation.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn left_padding_from(value: &[u8]) -> Address

Create a new byte array from the given slice src, left-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub fn right_padding_from(value: &[u8]) -> Address

Create a new byte array from the given slice src, right-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub const fn into_array(self) -> [u8; 20]

Returns the inner bytes array.

pub fn covers(&self, b: &Address) -> bool

Returns true if all bits set in b are also set in self.

pub const fn const_eq(&self, other: &Address) -> bool

Compile-time equality. NOT constant-time equality.

pub const fn bit_and(self, rhs: Address) -> Address

Computes the bitwise AND of two FixedBytes.

pub const fn bit_or(self, rhs: Address) -> Address

Computes the bitwise OR of two FixedBytes.

pub const fn bit_xor(self, rhs: Address) -> Address

Computes the bitwise XOR of two FixedBytes.

impl Address

pub fn from_word(word: FixedBytes<32>) -> Address

Creates an Ethereum address from an EVM word’s upper 20 bytes (word[12..]).

Examples

let word = b256!("0x000000000000000000000000d8da6bf26964af9d7eed9e03e53415d37aa96045");
assert_eq!(Address::from_word(word), address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045"));

pub fn into_word(&self) -> FixedBytes<32>

Left-pads the address to 32 bytes (EVM word size).

Examples

assert_eq!(
    address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045").into_word(),
    b256!("0x000000000000000000000000d8da6bf26964af9d7eed9e03e53415d37aa96045"),
);

pub fn parse_checksummed<S>( s: S, chain_id: Option<u64>, ) -> Result<Address, AddressError>
where S: AsRef<str>,

Parse an Ethereum address, verifying its EIP-55 checksum.

You can optionally specify an EIP-155 chain ID to check the address using EIP-1191.

Errors

This method returns an error if the provided string does not match the expected checksum.

Examples

let checksummed = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045";
let address = Address::parse_checksummed(checksummed, None).unwrap();
let expected = address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045");
assert_eq!(address, expected);

pub fn to_checksum(&self, chain_id: Option<u64>) -> String

Encodes an Ethereum address to its EIP-55 checksum into a heap-allocated string.

You can optionally specify an EIP-155 chain ID to encode the address using EIP-1191.

Examples

let address = address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045");

let checksummed: String = address.to_checksum(None);
assert_eq!(checksummed, "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045");

let checksummed: String = address.to_checksum(Some(1));
assert_eq!(checksummed, "0xD8Da6bf26964Af9d7EEd9e03e53415d37AA96045");

pub fn to_checksum_raw<'a>( &self, buf: &'a mut [u8], chain_id: Option<u64>, ) -> &'a mut str

Encodes an Ethereum address to its EIP-55 checksum into the given buffer.

For convenience, the buffer is returned as a &mut str, as the bytes are guaranteed to be valid UTF-8.

You can optionally specify an EIP-155 chain ID to encode the address using EIP-1191.

Panics

Panics if buf is not exactly 42 bytes long.

Examples

let address = address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045");
let mut buf = [0; 42];

let checksummed: &mut str = address.to_checksum_raw(&mut buf, None);
assert_eq!(checksummed, "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045");

let checksummed: &mut str = address.to_checksum_raw(&mut buf, Some(1));
assert_eq!(checksummed, "0xD8Da6bf26964Af9d7EEd9e03e53415d37AA96045");

pub fn to_checksum_buffer(&self, chain_id: Option<u64>) -> AddressChecksumBuffer

Encodes an Ethereum address to its EIP-55 checksum into a stack-allocated buffer.

You can optionally specify an EIP-155 chain ID to encode the address using EIP-1191.

Examples

let address = address!("0xd8da6bf26964af9d7eed9e03e53415d37aa96045");

let mut buffer: AddressChecksumBuffer = address.to_checksum_buffer(None);
assert_eq!(buffer.as_str(), "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045");

let checksummed: &str = buffer.format(&address, Some(1));
assert_eq!(checksummed, "0xD8Da6bf26964Af9d7EEd9e03e53415d37AA96045");

pub fn create(&self, nonce: u64) -> Address

Computes the create address for this address and nonce:

keccak256(rlp([sender, nonce]))[12:]

Examples

let sender = address!("0xb20a608c624Ca5003905aA834De7156C68b2E1d0");

let expected = address!("0x00000000219ab540356cBB839Cbe05303d7705Fa");
assert_eq!(sender.create(0), expected);

let expected = address!("0xe33c6e89e69d085897f98e92b06ebd541d1daa99");
assert_eq!(sender.create(1), expected);

pub fn create2_from_code<S, C>(&self, salt: S, init_code: C) -> Address
where S: Borrow<[u8; 32]>, C: AsRef<[u8]>,

Computes the CREATE2 address of a smart contract as specified in EIP-1014:

keccak256(0xff ++ address ++ salt ++ keccak256(init_code))[12:]

The init_code is the code that, when executed, produces the runtime bytecode that will be placed into the state, and which typically is used by high level languages to implement a ‘constructor’.

Examples

let address = address!("0x8ba1f109551bD432803012645Ac136ddd64DBA72");
let salt = b256!("0x7c5ea36004851c764c44143b1dcb59679b11c9a68e5f41497f6cf3d480715331");
let init_code = bytes!("6394198df16000526103ff60206004601c335afa6040516060f3");
let expected = address!("0x533ae9d683B10C02EbDb05471642F85230071FC3");
assert_eq!(address.create2_from_code(salt, init_code), expected);

pub fn create2<S, H>(&self, salt: S, init_code_hash: H) -> Address
where S: Borrow<[u8; 32]>, H: Borrow<[u8; 32]>,

Computes the CREATE2 address of a smart contract as specified in EIP-1014, taking the pre-computed hash of the init code as input:

keccak256(0xff ++ address ++ salt ++ init_code_hash)[12:]

The init_code is the code that, when executed, produces the runtime bytecode that will be placed into the state, and which typically is used by high level languages to implement a ‘constructor’.

Examples

let address = address!("0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f");
let salt = b256!("0x2b2f5776e38002e0c013d0d89828fdb06fee595ea2d5ed4b194e3883e823e350");
let init_code_hash =
    b256!("0x96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f");
let expected = address!("0x0d4a11d5EEaaC28EC3F61d100daF4d40471f1852");
assert_eq!(address.create2(salt, init_code_hash), expected);

pub fn create_eof<S>(&self, salt: S) -> Address
where S: Borrow<[u8; 32]>,

Computes the address created by the EOFCREATE opcode, where self is the sender.

The address is calculated as keccak256(0xff || sender32 || salt)[12:], where sender32 is the sender address left-padded to 32 bytes with zeros.

See EIP-7620 for more details.

This function's stability is not guaranteed. It may change in the future as the EIP is not yet accepted.

Examples

let address = address!("0xb20a608c624Ca5003905aA834De7156C68b2E1d0");
let salt = b256!("0x7c5ea36004851c764c44143b1dcb59679b11c9a68e5f41497f6cf3d480715331");
// Create an address using CREATE_EOF
let eof_address = address.create_eof(salt);

pub fn from_raw_public_key(pubkey: &[u8]) -> Address

Instantiate by hashing public key bytes.

Panics

If the input is not exactly 64 bytes

pub fn from_public_key(pubkey: &VerifyingKey<Secp256k1>) -> Address

Converts an ECDSA verifying key to its corresponding Ethereum address.

pub fn from_private_key(private_key: &SigningKey<Secp256k1>) -> Address

Converts an ECDSA signing key to its corresponding Ethereum address.