hopr_chain_types::utils

Function fund_channel

pub async fn fund_channel<M: Middleware>(
    counterparty: Address,
    hopr_token: HoprToken<M>,
    hopr_channels: HoprChannels<M>,
    amount: U256,
)

Funds the channel to the counterparty with the given amount of HOPR tokens. The amount must be present in the wallet of the client.