Crate hopr_chain_types

This crate contains various on-chain related modules and types.

Modules

chain_events

Types that are related to events that are raised on-chain and extracted from chain logs.

errors

exports

payload

Module defining various Ethereum transaction payload generators for the actions.

prelude

Structs

AlloyAddress

An Ethereum address, 20 bytes in length.

ContractAddresses

Holds addresses of all smart contracts.

Enums

ParsedHoprChainAction

Represents the action previously parsed from an EIP-2718 transaction.

Functions

contract_addresses_for_network

Returns chain ID and contract addresses for a known HOPR network on-chain deployment with the given name (e.g. rotsee or dufour)